% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{as__degminsec}
\alias{as__degminsec}
\alias{as__degminsec.numeric}
\alias{as__degmin}
\alias{as__degmin.numeric}
\alias{as__decdeg}
\alias{as__decdeg.numeric}
\title{Convert Coordinate to Degrees, Minutes and Seconds}
\usage{
as__degminsec(object, ...)

\method{as__degminsec}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)

as__degmin(object, ...)

\method{as__degmin}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)

as__decdeg(object, ...)

\method{as__decdeg}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)
}
\arguments{
\item{object}{a \code{"coord"} object or a \code{numeric} vector.}

\item{...}{further arguments passed to or from other methods.}

\item{.fmt}{\code{character} string indicating the format of \code{object}; must be one of
\code{"decdeg"} (default), \code{"degmin"} or \code{"degminsec"}.}

\item{.as_numeric}{logical, signifying whether to return a \code{numeric} value, if \code{TRUE}, or a
\code{"coord"} object, if \code{FALSE}, the default.}
}
\value{
A \code{\link[=coord]{"coord"}} object or \code{numeric} value in the desired format.
}
\description{
Convert the format of a geographic or GPS coordinate to degrees, minutes and seconds.
}
\details{
Converts between coordinates represented in decimal degrees ("decdeg"), integer degrees and
decimal minutes ("degmin"), and integer degrees, integer minutes, and decimal seconds
("degminsec"). Works with individual \code{\link[=coord]{"coord"}} objects returned using the
\code{\link[=coord]{coord()}} function, or with vectors of simple numeric values.
See \code{\link[=coord]{"coord"}} for further details.
}
\examples{
## `"coord"` objects in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(coord_dd <- coord(51.507765, "decdeg"))
(coord_dm <- coord(5130.4659, "degmin") |> as__degminsec())
(coord_dms <- coord(513027.95, "degminsec") |> as__degmin())

## as__degminsec
coord_dd |> as__degminsec()
coord_dm |> as__degminsec()

## as__degmin
coord_dd |> as__degmin()
coord_dms |> as__degmin()

## as__decdeg
coord_dm |> as__decdeg()
coord_dms |> as__decdeg()

## Numeric vectors in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(num_dd <- c(51.507765, 49.546210, 48.107232, 38.889494, 0.000000, -37.111740, -53.104781))
(num_dm <- c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869))
(num_dms <- c(513027.95, 493246.36, 480626.04, 385322.18, 0.00, -370642.26, -530617.21))

## as__degminsec
num_dd |> as__degminsec(.fmt = "decdeg")
num_dm |> as__degminsec(.fmt = "degmin")

num_dd |> as__degminsec(.fmt = "decdeg", .as_numeric = TRUE)
num_dm |> as__degminsec(.fmt = "degmin", .as_numeric = TRUE)

## as__degmin
num_dd |> as__degmin(.fmt = "decdeg")
num_dms |> as__degmin(.fmt = "degminsec")

num_dd |> as__degmin(.fmt = "decdeg", .as_numeric = TRUE)
num_dms |> as__degmin(.fmt = "degminsec", .as_numeric = TRUE)

## as__decdeg
num_dm |> as__decdeg(.fmt = "degmin")
num_dms |> as__decdeg(.fmt = "degminsec")

num_dm |> as__decdeg(.fmt = "degmin", .as_numeric = TRUE)
num_dms |> as__decdeg(.fmt = "degminsec", .as_numeric = TRUE)

rm(coord_dd, coord_dm, coord_dms, num_dd, num_dm, num_dms)
}
\seealso{
Other coord: 
\code{\link{coord}()}
}
\concept{coord}
