% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{as_coord}
\alias{as_coord}
\alias{as_coord.coord}
\alias{as_coord.numeric}
\title{Convert Coordinate or Numeric to Another Coordinate Format}
\usage{
as_coord(object, ...)

\method{as_coord}{coord}(object, ..., .fmt = c("decdeg", "degmin", "degminsec"))

\method{as_coord}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .latorlon = c(NA, "lat", "lon"),
  .as_numeric = FALSE
)
}
\arguments{
\item{object}{a \code{"coord"} object or a \code{numeric} vector.}

\item{...}{further arguments passed to or from other methods.}

\item{.fmt}{\code{character} string indicating the format of \code{object}; must be one of \code{"decdeg"}
(default), \code{"degmin"} or \code{"degminsec"}.}

\item{.as_numeric}{logical, signifying whether to return a \code{"coord"} object or a \code{numeric} value;
default \code{FALSE}, the former.}
}
\value{
A \code{\link[=coord]{"coord"}} object or \code{numeric} value in the desired format.
}
\description{
Convert the format of geographic or GPS coordinates represented by \code{"coord"} objects or \code{numeric}
values among decimal degrees, degrees and minutes, and degrees, minutes and seconds.
}
\details{
Converts between coordinates represented in decimal degrees ("decdeg"), integer degrees and
decimal minutes ("degmin"), and integer degrees, integer minutes, and decimal seconds
("degminsec"). Works with individual \code{\link[=coord]{"coord"}} objects returned using the
\code{\link[=coord]{coord()}} function, or with vectors of simple numeric values.
See \code{\link[=coord]{"coord"}} for further details.
}
\examples{
## `"coord"` objects in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(coord_dd <- coord(51.507765, "decdeg"))
(coord_dm <- coord(5130.4659, "degmin"))
(coord_dms <- coord(513027.95, "degminsec"))
}
\seealso{
Other coord: 
\code{\link{coord}()}
}
\concept{coord}
