% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{as_degminsec}
\alias{as_degminsec}
\alias{as_degminsec.numeric}
\alias{as_degmin}
\alias{as_degmin.numeric}
\alias{as_decdeg}
\alias{as_decdeg.numeric}
\title{Convert Coordinate Format}
\usage{
as_degminsec(object, ...)

\method{as_degminsec}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)

as_degmin(object, ...)

\method{as_degmin}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)

as_decdeg(object, ...)

\method{as_decdeg}{numeric}(
  object,
  ...,
  .fmt = c("decdeg", "degmin", "degminsec"),
  .as_numeric = FALSE
)
}
\arguments{
\item{object}{a \code{"coord"} object or a \code{numeric} vector.}

\item{...}{further arguments passed to or from other methods.}

\item{.fmt}{\code{character} string indicating the format of \code{object}; must be one of \code{"decdeg"}
(default), \code{"degmin"} or \code{"degminsec"}.}

\item{.as_numeric}{logical, signifying whether to return a \code{"coord"} object or a \code{numeric} value;
default \code{FALSE}, the former.}
}
\value{
A \code{\link[=coord]{"coord"}} object or \code{numeric} value in the desired format.
}
\description{
Convert the format of geographic or GPS coordinates to decimal degrees, to degrees and minutes
or to degrees, minutes and seconds.
}
\details{
Converts between coordinates represented in decimal degrees ("decdeg"), integer degrees and
decimal minutes ("degmin"), and integer degrees, integer minutes, and decimal seconds
("degminsec"). Works with individual \code{\link[=coord]{"coord"}} objects returned using the
\code{\link[=coord]{coord()}} function, or with vectors of simple numeric values.
See \code{\link[=coord]{"coord"}} for further details.
}
\examples{
## `"coord"` objects in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(coord_dd <- coord(51.507765, "decdeg"))
(coord_dm <- coord(5130.4659, "degmin"))
(coord_dms <- coord(513027.95, "degminsec"))

## as_degminsec
coord_dd |> as_degminsec()
coord_dm |> as_degminsec()

## as_degmin
coord_dd |> as_degmin()
coord_dms |> as_degmin()

## as_decdeg
coord_dm |> as_decdeg()
coord_dms |> as_decdeg()

## `"latnlon"` objects in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(latnlon_dd <- coord(c(51.507765, -0.127924), "decdeg", "both"))
(latnlon_dm <- coord(c(5130.4659, -7.6754), "degmin", "both"))
(latnlon_dms <- coord(c(513027.95, -740.53), "degminsec", "both"))

## as_degminsec
latnlon_dd |> as_degminsec()
latnlon_dm |> as_degminsec()

## as_degmin
latnlon_dd |> as_degmin()
latnlon_dms |> as_degmin()

## as_decdeg
latnlon_dm |> as_decdeg()
latnlon_dms |> as_decdeg()

## Numeric vectors in decimal degrees; in degrees and minutes;
##   and in degrees, minutes, and seconds
(num_dd <- c(51.507765, 49.546210, 48.107232, 38.889494, 0.000000, -37.111740, -53.104781))
(num_dm <- c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869))
(num_dms <- c(513027.95, 493246.36, 480626.04, 385322.18, 0.00, -370642.26, -530617.21))

## as_degminsec
num_dd |> as_degminsec(.fmt = "decdeg")
num_dm |> as_degminsec(.fmt = "degmin")

num_dd |> as_degminsec(.fmt = "decdeg", .as_numeric = TRUE)
num_dm |> as_degminsec(.fmt = "degmin", .as_numeric = TRUE)

## as_degmin
num_dd |> as_degmin(.fmt = "decdeg")
num_dms |> as_degmin(.fmt = "degminsec")

num_dd |> as_degmin(.fmt = "decdeg", .as_numeric = TRUE)
num_dms |> as_degmin(.fmt = "degminsec", .as_numeric = TRUE)

## as_decdeg
num_dm |> as_decdeg(.fmt = "degmin")
num_dms |> as_decdeg(.fmt = "degminsec")

num_dm |> as_decdeg(.fmt = "degmin", .as_numeric = TRUE)
num_dms |> as_decdeg(.fmt = "degminsec", .as_numeric = TRUE)

rm(coord_dd, coord_dm, coord_dms, num_dd, num_dm, num_dms)
}
\seealso{
Other coord: 
\code{\link{coord}()}
}
\concept{coord}
