% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BitsnBobs-package.R
\docType{package}
\name{BitsnBobs-package}
\alias{BitsnBobs}
\alias{BitsnBobs-package}
\title{BitsnBobs: For General Bits and Bobs of Code}
\description{
Does this and that using functions: as_rostido(), boxcox3(), browse(), cat_names(), cc_rate(), cohens_kappa(), const(), cor_coef.test(), count_lgl(), data_wizard(), decdeg(), decdeg_to_dms(), degminsec(), design_effect(), detective(), detective()<-, dms_to_decdeg(), eff_rate(), endstop(), endstop_data(), facet_histo(), fct_to_num(), formul_pwrseq(), get_amortint(), get_amortval(), iterate(), j(), j2eff_rate(), known_s3generics(), kwd_cols(), kurtosis(), kwd_cols()<-, latlon(), latlon_dd(), lf(), lgl_cols(), list_lgl(), logit(), ls_all(), marker(), method_info(), most_recent_fname(), nom_rate(), opt_bc(), phi_coef(), phi_coef.test(), power_seq(), print_all(), print_lf(), prob_from_logit(), read_triodos_csv(), recursive(), remplacer(), retriever(), revmat(), rm_objects(), s3_in_namespace(), s3debug(), s3mag7(), sample_size(), skew(), skew.test(), skewness(), skewness.test(), split_to_cols(), split_to_rows(), starsig(), sum_lgl(), triodos_fname(), wizard(); infix operators: op-min-max; and datasets: heights, litter_sizes, starwars2 and starwars3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mark-eis.github.io/BitsnBobs/}
}

}
\author{
\strong{Maintainer}: Mark Eisler \email{Mark.Eisler@bristol.ac.uk} (\href{https://orcid.org/0000-0001-6843-3345}{ORCID}) [copyright holder]

}
\keyword{internal}
