% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catnames.R
\name{cat_names}
\alias{cat_names}
\title{Concatenate First Names and Surnames in a Data Frame into Delimiter-Separated Character Strings}
\usage{
cat_names(
  data,
  .firstname = Firstname,
  .surname = Surname,
  ...,
  .delimiter = ", "
)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a tibble).}

\item{.firstname}{quoted name of column containing first names; default \code{Firstname}.}

\item{.surname}{quoted name of column containing surnames; default \code{Surname}.}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> names of variables to group by.}

\item{.delimiter}{a character string to separate the first name-surname pairs, see \code{collapse} argument of
\code{\link[base:paste]{paste}}; default \code{", "}.}
}
\value{
A \link[tibble:tibble-package]{tibble} data frame containing the column \code{Names} and further individual columns
for each grouping variable.
}
\description{
Concatenates first names and surnames contained in individual columns of a data frame. The resulting first name-surname
pairs are themselves further combined into one or more delimiter-separated character strings on the basis of selected
grouping variables.
}
\details{
A number of first name-surname pairs in \code{data} are all concatenated in a single string for each of one or more
grouping variables selected using \code{\dots}. Grouping variables in \code{.data} are selected using the \code{\dots}
argument with the <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax of package \pkg{dplyr}, including use of
\strong{selection helpers}.

By default, the name pairs within a string are separated using commas or otherwise using \code{.delimiter}, if
provided.
}
\examples{

starwars2 ## Has name column of original starwars split into Firstname and Surname

starwars2 |>
  cat_names(Firstname, Surname, homeworld) |>
  print_all()

starwars2 |>
  cat_names(,, species, .delimiter = "; ") |>
  print_all()

starwars2 |>
  cat_names(,, homeworld, species) |>
  print_all()

}
\seealso{
\code{\link[base:paste]{paste}}, <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>.
}
