% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{catapult}
\alias{catapult}
\alias{print.catapult}
\title{Catapult Class for Consistent Printing}
\usage{
catapult(object = vector(), lead = "Catapult", revert = FALSE, ...)

\method{print}{catapult}(x, ...)
}
\arguments{
\item{object}{Object to be converted to \code{"catapult"} class.}

\item{lead}{a \code{character} string giving the title to be printed.}

\item{revert}{\code{logical} indicating whether the \code{"revert"} attribute should be set to \code{TRUE} or \code{FALSE}. By default,
when \code{FALSE}, the corresponding print method \code{print.catapult()} will invoke further print methods for the underlying
inherited \code{object}; if \code{TRUE}, \code{print.catapult()} will return to the calling function after printing \code{lead}.}

\item{\dots}{additional named arguments to be forwarded to print methods of classes inherited from \code{object}, for
example as required for formatting and printing a \code{\link[base:print.dataframe]{data frame}}.}

\item{x}{an object used to select a method.}
}
\value{
An object of class \code{"catapult"} inheriting class(es) from \code{object}.
}
\description{
Creates an object of class \code{"catapult"} with a built-in title string used for printing.
}
\details{
\code{catapult()} converts an object to class \code{"catapult"}, inheriting from its existing class(es).
}
\note{
Deprecated, use instead \pkg{\link[ParaAnita]{ParaAnita}} \code{\link[ParaAnita:announce]{announce()}}.
}
\examples{
catapult()
(cpt <- catapult("x", lead = "Lorem ipsum dolor sit amet"))
.class2(cpt)
attr(cpt, "revert")

(cpt <- catapult("x", lead = "Lorem ipsum dolor sit amet", TRUE))
.class2(cpt)
attr(cpt, "revert")

rm(cpt)
}
\seealso{
\code{\link[base:cat]{cat()}}, \code{\link[base:print]{print()}}.

Other print: 
\code{\link{lf}()},
\code{\link{print_all}()}
}
\concept{print}
\keyword{print}
