% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-cols.R
\name{count_lgl}
\alias{count_lgl}
\alias{sum_lgl}
\title{Count Combinations and Totals for Logical Columns in Data Frame}
\usage{
count_lgl(df, .newcol = n, .arrange_by = NULL)

sum_lgl(df, wt = NULL)
}
\arguments{
\item{df}{a data frame, or a data frame extension (e.g. a tibble).}

\item{.newcol}{quoted name to be assigned to the new count column; default \code{n}.}

\item{.arrange_by}{<\code{\link[rlang:args_data_masking]{data-masked}}> quoted name of a column for ordering results; default
\code{NULL}.}

\item{wt}{frequency weights. Can be \code{NULL} or the name of a numeric variable column; default \code{NULL}.
\itemize{
\item If \code{NULL} (the default), counts the number of rows with \code{TRUE} values for each logical column.
\item If a variable, computes sum(wt) for rows with \code{TRUE} values for each logical column.
}}
}
\value{
\code{count_lgl()} returns a data frame comprising unique combinations of the logical columns of \code{df}
and an additional column of the counts of ocurrences of each of these combinations. \code{sum_lgl()} returns a named
numeric vector of totals for each logical column.
}
\description{
Functions to count the number of ocurrences of unique combinations of values across all logical columns in a data
frame and the total number of \code{TRUE} values for each logical column.
}
\details{
Function \code{count_lgl()} counts the number of ocurrences of unique combinations of values across all logical
columns in a data frame using \code{\link[dplyr:summarise]{summarise}} in package \pkg{dplyr}.

Function \code{sum_lgl()} counts the total number of \code{TRUE} values for each logical column in a data frame using
matrix multiplication.

These functions may be useful for counting combinations or tallying totals of keywords flagged as \code{TRUE} in
logical columns added to a data frame using \code{\link{kwd_cols}}.
}
\examples{
## Following on from kwd_cols() examples… 
car_names <- data.frame(Response = rownames(mtcars)) |>
  kwd_cols(, c("Mazda", "Merc", "Toyota", "Volkswagen", "X", "450"))
car_names

(carname_counts <- count_lgl(car_names, .arrange_by = desc(n)))
sum_lgl(car_names)
sum_lgl(carname_counts, wt = "n")

car_names |> count_lgl(.newcol = subtotals) |> print_lf() |> sum_lgl(wt = "subtotals")

rm(car_names, carname_counts)
}
\seealso{
\code{\link[dplyr:summarise]{summarise}}.

Other logical-cols: 
\code{\link{kwd_cols}()},
\code{\link{list_lgl}()}
}
\concept{logical-cols}
