% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{decdeg_to_dms}
\alias{decdeg_to_dms}
\alias{decdeg_to_dms.default}
\alias{decdeg_to_dms.decdeg}
\alias{decdeg_to_dms.list}
\title{Convert Decimal Degrees to Degrees, Minutes and Seconds}
\usage{
decdeg_to_dms(object, ...)

\method{decdeg_to_dms}{default}(object, ...)

\method{decdeg_to_dms}{decdeg}(object, ...)

\method{decdeg_to_dms}{list}(object, ...)
}
\arguments{
\item{object}{a numeric or an object of class \code{\link[=decdeg]{"decdeg"}}, representing a coordinate of latitude or
longitude in decimal degrees or a list of \code{"decdeg"} objects.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{"degminsec"}, or if \code{length(object) > 1}, a \code{list} of such objects, representing a
coordinate of latitude or longitude in degrees, minutes and seconds as a named list with components: -

\item{deg}{degrees represented by an integer with maximum absolute value of 180.}

\item{min}{minutes represented by a positive integer with value less than 60.}

\item{sec}{seconds represented by a positive numeric with value less than 60.}
}
\description{
Convert decimal degrees to degrees, minutes and seconds.
}
\details{
\code{decdeg_to_dms()} is an S3 function that works with individual coordinates, latitude and longitude
values paired in a named list (see examples) or with a longer list of coordinates.
}
\examples{
decdeg_to_dms(49.54621)

(coord <- decdeg(49.54621))
decdeg_to_dms(coord)

(coords <- decdeg(c(lat = 49.54621, lon = 18.398562)))
decdeg_to_dms(coords)

rm(coord, coords)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{degminsec}()},
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
