% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{design_effect}
\alias{design_effect}
\title{Design Effect for Cluster Sampling and Cluster Randomised Trials}
\usage{
design_effect(m, ri)
}
\arguments{
\item{m}{integer representing cluster size.}

\item{ri}{intracluster (=intraclass) correlation coefficient.}
}
\value{
Numeric value of design effect.
}
\description{
The design effect is the ratio of the total number of subjects required using cluster randomisation to the number
required using individual randomisation.
}
\details{
The design effect can be presented neatly in terms of the intracluster correlation and the number in a single cluster: -

\deqn{D = 1 + (m − 1)r_{I}}{%
       D = 1 + (m − 1)r_{I}}

If there is only one observation per cluster, \code{m = 1}, the design effect is 1.0 and the two designs are the same.
Otherwise, the larger the intracluster correlation — that is, the more important the variation between clusters is — the
bigger the design effect and the more subjects we will need to get the same power as a simply randomised study. Even a
small intracluster correlation will have an impact if the cluster size is large.
}
\note{
\emph{Description} and \emph{Details} taken verbatim from the second reference.
}
\examples{
## Example: x-ray guidelines study from second reference.
design_effect(m = 50L, ri = 0.019)

}
\references{
Kerry, S.M. & Bland, J.M., 1998. Sample size in cluster randomisation. \emph{Brit Med J} \strong{316}: 5490.
\href{https://doi.org/10.1136/bmj.316.7130.549}{\doi{10.1136/bmj.316.7130.549}}.

Kerry, S.M. & Bland, J.M., 1998. The intracluster correlation coefficient in cluster randomisation. \emph{Brit Med J}
\strong{316}: 1455-1460. \href{https://doi.org/10.1136/bmj.316.7142.1455}{\doi{10.1136/bmj.316.7142.1455}}.
}
\seealso{
Other sample-size: 
\code{\link{sample_size}()}
}
\concept{sample-size}
\keyword{cluster}
\keyword{survey}
