% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{dms_to_decdeg}
\alias{dms_to_decdeg}
\alias{dms_to_decdeg.default}
\alias{dms_to_decdeg.degminsec}
\alias{dms_to_decdeg.list}
\alias{dms_to_decdeg.latlon}
\title{Convert Degrees, Minutes and Seconds to Decimal Degrees}
\usage{
dms_to_decdeg(object, ...)

\method{dms_to_decdeg}{default}(object, ..., .after = c("deg", "min", "sec"))

\method{dms_to_decdeg}{degminsec}(object, ...)

\method{dms_to_decdeg}{list}(object, ...)

\method{dms_to_decdeg}{latlon}(object, ...)
}
\arguments{
\item{object}{a \code{numeric} or an object of class \code{\link[=degminsec]{"degminsec"}}, representing a coordinate of latitude
or longitude in degrees, minutes and seconds or a list of \code{"degminsec"} objects.}

\item{...}{further arguments passed to or from other methods.}

\item{.after}{a \code{character} string indicating the position of the decimal point in \code{object}; must be one of
\code{"deg"} (default), \code{"min"}, or \code{"sec"}. You can specify just the initial letter.}
}
\value{
An object of class \code{"decdeg"}, or if \code{length(object) > 1}, a list of such objects, instantiating a
coordinate of latitude or longitude in decimal degrees represented by a numeric of type \code{double} with maximum
absolute value of \var{180}˚. Attribute \code{".latorlon"} indicates whether the object is a coordinate of latitude
or longitude.
}
\description{
Convert degrees, minutes and seconds to decimal degrees.
}
\details{
\code{dms_to_decdeg()} is an S3 function that works with individual coordinates supplied as \verb{numeric values} or as
\code{\link[=degminsec]{"degminsec"}} objects, or with lists of such coordinates.  It also accepts, latitude and
longitude values paired in a \code{\link[=latlon]{"latlon"}} object (see examples), or lists of \code{"latlon"}
paired coordinates.
}
\examples{
dms_to_decdeg(49.3246368)
dms_to_decdeg(4932.46368, .after = "min")
dms_to_decdeg(493246.368, .after = "sec")

(coord <- degminsec(49.3246368))
dms_to_decdeg(coord)

(coords <- latlon(c(49.3246368, 18.2354822)))
dms_to_decdeg(coords)

rm(coord, coords)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()},
\code{\link{latlon}()}
}
\concept{degreeconvert}
\keyword{utilities}
