% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage.R
\name{eff_rate}
\alias{eff_rate}
\alias{nom_rate}
\alias{cc_rate}
\title{Effective Rate of Interest from a Nominal Rate and its Conversion Frequency or the Inverse}
\usage{
eff_rate(nom_rate, con_fq)

nom_rate(eff_rate, con_fq)

cc_rate(eff_rate)
}
\arguments{
\item{nom_rate}{numeric, the nominal interest rate convertible \code{con_fq} times per year.}

\item{con_fq, }{integer, the interest conversion frequency per year.}

\item{eff_rate}{numeric, the effective rate of interest.}
}
\value{
A numeric: \code{eff_rate()} returns the effective rate of interest; \code{nom_rate()} returns the nominal
interest rate convertible \var{con_fq} times per year.
}
\description{
\code{eff_rate()} calculates the effective rate of annual interest from a nominal rate and its interest conversion
frequency.

\code{nom_rate()} calculates the nominal rate of interest for a given interest conversion frequency from an effective rate.

\code{cc_rate()} calculates the continuous compounding rate of interest from an effective rate.
}
\details{
Effective Rate of Interest (er): —

\deqn{er = \displaystyle \left(1 + \frac{nr}{ic} \right)^{ic} - 1}{%
    er = (1 + nr/ic)<sup>ic</sup> - 1}

Nominal Rate of Interest (nr): —

\deqn{nr = \displaystyle \left(\left(1 + er \right)^{\frac{1}{ic}} - 1 \right).ic}{%
    nr = ((1 + er)\<sup>1/ic</sup> - 1).ic}

Continuous Compounding Rate of Interest (ccr): —

\deqn{ccr = \displaystyle e^{er} - 1}{%
    ccr = exp(er) - 1}

where \eqn{er} is the effective rate of interest, \eqn{nr} is the nominal rate of interest, and \eqn{ic} is
the interest conversion frequency. See \code{\link[FinancialMath:amort.period]{amort.period()}} in package
\CRANpkg{FinancialMath} for further details.
}
\examples{
## A lender calculates interest daily using a simple annual rate of 5.7480\%, based on a 365-day year:
## Calculate effective rate: -
eff_rate(0.057480, 365)

## Calculate nominal rates at monthly and daily interest conversion frequencies: -
## Monthly
nom_rate(0.05915929, 12)
## Daily—same as the original simple annual rate
nom_rate(0.05915929, 365)

## Example from https://www.investopedia.com/terms/e/effectiveinterest.asp
eff_rate(0.10, c(2, 4, 12, 365)) |>
    setNames(c("Semiannual", "Quarterly", "Monthly", "Daily"))

## Limit of compounding (ibid.)
eff_rate(0.1, c(365, 365 * 24, 365 * 24 * 60, 365 * 24 * 60 * 60)) |>
    setNames(c("Daily", "Hourly", "Minutely", "Secondly"))
## Continuous compounding
cc_rate(0.1)
}
\references{
Investopedia: Effective Annual Interest Rate: Definition, Formula, and Example.
\href{https://www.investopedia.com/terms/e/effectiveinterest.asp}{www.investopedia.com/terms/e/effectiveinterest}

Investopedia: Continuous Compounding Definition and Formula.
\href{https://www.investopedia.com/terms/c/continuouscompounding.asp}{
www.investopedia.com/terms/c/continuouscompounding}
}
\seealso{
\code{\link[FinancialMath:amort.period]{amort.period()}}, \code{\link[FinancialMath:amort.table]{amort.table()}}

Other amort: 
\code{\link{get_amortval}()},
\code{\link{j}()}
}
\concept{amort}
