% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{endstop}
\alias{endstop}
\alias{endstop_data}
\title{End String with Full Stop and no Other Punctuation or Spaces.}
\usage{
endstop(string, .stop = TRUE)

endstop_data(.data, ..., .stop = TRUE)
}
\arguments{
\item{string}{a character vector of length one.}

\item{.stop}{\code{logical}. Whether or not to add a full stop at the end of the string; default \code{TRUE}.}

\item{.data}{a data frame, or a data frame extension (e.g. a \code{\link[tibble:tibble-package]{tibble}}).}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> character columns in \var{.data} to \code{endstop}.}
}
\value{
For \code{endstop()}, a character vector of length one, optionally terminating in a full stop. For
\code{endstop_data()}, a \code{\link[tibble:tibble-package]{tibble}} derived from \code{.data}, with selected character
columns modified by \code{endstop()}.
}
\description{
\code{endstop()} removes all punctuation and spaces from the end of a string and optionally terminates the
string with a full stop.

\code{endstop_data()} removes all punctuation and spaces from the end of selected strings in \var{.data} and
optionally terminates the strings with full stops.
}
\details{
Uses \code{\link[stringr:str_detect]{str_detect}} from package \pkg{\link[stringr]{stringr}} to detect the regular
expressions \code{'[:punct:]'} and \code{'[:space:]'}, and \code{\link[stringr:str_sub]{str_sub}} to modify the string.

For \code{endstop_data()}, character columns in \code{.data} are selected using \code{\dots} with the
<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax of package \link{dplyr}, including use of
\strong{selection helpers} and modified by \code{endstop()}. If no character columns are selected in \code{\dots},
all character columns in \var{.data} will be modified  by \code{endstop()}.
}
\examples{
"Mimiland" |> endstop()
"Mimiland." |> endstop(FALSE)
"Mimiland," |> endstop()
"Mimiland ." |> endstop()
"Mimiland. " |> endstop()
s <- "Mimiland.!?\\\\(){}"
cat(s)
endstop(s)
s <- "Mimiland . ! ? \\\\ ( ) { } "
cat(s)
endstop(s, FALSE)

starwars3 |> endstop_data(name)
starwars3 |> endstop_data(starts_with("sk"))
starwars3 |> endstop_data()

rm(s)
}
\seealso{
\code{\link[stringr:str_detect]{str_detect()}}, \code{\link[stringr:str_sub]{str_sub()}}

Other utils: 
\code{\link{const}()},
\code{\link{marker}()},
\code{\link{op-min-max}},
\code{\link{revmat}()}
}
\concept{utils}
