% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BitsnBobs.R
\name{fct_to_num}
\alias{fct_to_num}
\title{Factor As Numeric}
\usage{
fct_to_num(f)
}
\arguments{
\item{f}{factor to be converted to numeric values}
}
\value{
Numeric
}
\description{
Transform a factor to approximately its original numeric values.
}
\details{
See \sQuote{Warning} section of \code{\link[base:factor]{factor}}: \enc{–}{-}

\dQuote{In particular, \code{as.numeric} applied to a factor is meaningless, and may happen by implicit
coercion. To transform a factor \code{f} to approximately its original numeric values,
\code{as.numeric(levels(f))[f]} is recommended and slightly more efficient than
\code{as.numeric(as.character(f))}.}

Accordingly, \code{fct_to_num()} implements this method.
}
\examples{
## Create sequence of numeric values as a factor
(f <- factor(2001:2020))

## Seemingly numeric levels are actually character strings!
levels(f) |> str()

## Returns codes for factor levels, not what was expected
f |> as.numeric()

## Returns numeric values equivalent to factor levels
f |> fct_to_num()

rm(f)
}
\seealso{
\code{\link[base:factor]{factor}}
}
\keyword{category}
\keyword{math}
