% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerseq.R
\name{formul_pwrseq}
\alias{formul_pwrseq}
\title{Expand Term In Right Hand Side Of A Formula As A Power Sequence}
\usage{
formul_pwrseq(base_fla, n, ...)
}
\arguments{
\item{base_fla}{the a term in the formula to be expanded as a power sequence.}

\item{n}{a non-negative integer or number coercible into a positive integer.}

\item{...}{additional terms in the formula.}
}
\value{
A \code{formula}.
}
\description{
Create a power sequence from a term in a \code{\link[stats]{formula}}, \var{base_fla}, and the maximum power number,
\var{n}, optionally including other formula terms.
}
\examples{
formul_pwrseq(y ~ x, 5)
formul_pwrseq(y ~ log(x), 5)
formul_pwrseq(y ~ A, 3, B, C, D)
}
\seealso{
\code{\link[stats]{formula}}.

Other powerseq: 
\code{\link{power_seq}()}
}
\concept{powerseq}
