% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage.R
\name{get_amortval}
\alias{get_amortval}
\alias{get_amortint}
\title{Individual Values from the Output of amort.period or amort.table()}
\usage{
get_amortval(x, val = "Eff Rate")

get_amortint(x, until)
}
\arguments{
\item{x}{either a matrix of input variables, calculated unknown variables, and amortization figures for a given
period, as  output by \code{amort.period()} or a list of two components, \code{Schedule} (a data frame of an amortization
schedule) and \code{Other}, (a matrix of the input variables and other calculated variables) as output by \code{amort.table()}.}

\item{val}{a \code{character} string giving the exact name of the value to be returned.}

\item{until}{a \code{numeric} signifying the period up to and including which total interest should be summated.}
}
\value{
A numeric value.
}
\description{
Functions to return individual values from the output of \code{amort.period()} or \code{amort.table()} in package
\pkg{FinancialMath}.
}
\details{
\code{get_amortval()} returns individual values from the output of \code{\link[FinancialMath:amort.period]{amort.period()}} or
the \code{Other} list element of the output of \code{\link[FinancialMath:amort.table]{amort.table()}} in package
\CRANpkg{FinancialMath}.

\code{get_amortint()} returns the total value of interest payments from the \code{Schedule} list element of the output of
\code{\link[FinancialMath:amort.table]{amort.table()}}, up to and including the period given in \code{until}.
}
\examples{
\dontshow{
## Requires {FinancialMath} package
if (!requireNamespace("FinancialMath", quietly = TRUE)) 
  warning("package 'FinancialMath' must be installed")
}
(apd <- try(
    FinancialMath::amort.period(Loan = 200099, n = 15 * 12, i = 0.0184, ic = 12, pf = 12, t = 60)
))

## Default value is "Eff Rate"
get_amortval(apd)
get_amortval(apd, "PMT")  
get_amortval(apd, "Balance")  

atb <- try(FinancialMath::amort.table(Loan = 200099, n = 15 * 12, i = 0.0184, ic = 12, pf = 12))
atb$Other

## Default value is "Eff Rate"
get_amortval(atb)
get_amortval(atb, "Total Paid")
get_amortval(atb, "Total Interest")

get_amortint(atb, 60)

rm(apd, atb)
}
\seealso{
\code{\link[FinancialMath:amort.period]{amort.period()}}, \code{\link[FinancialMath:amort.table]{amort.table()}}

Other amort: 
\code{\link{eff_rate}()},
\code{\link{j}()}
}
\concept{amort}
