% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heights}
\alias{heights}
\title{University Student Heights}
\format{
\subsection{}{

An \code{integer vector} with length \code{100}.
\describe{
\item{heights}{Height of the student in inches.}
}
}
}
\source{
Adapted from Spiegel, Murray R., and Larry J. Stephens, (2008). \emph{Theory and Problems of Statistics}.
4th edn. McGraw-Hill.
\href{http://www.buders.com/UNIVERSITE/Universite_Dersleri/istatistik/statistics.pdf}{ISBN:9780071755498}

See also: \href{https://brownmath.com/stat/shape.htm#Example1}{https://brownmath.com}
}
\usage{
heights
}
\description{
Heights of 100 randomly selected male university students from Table 2.1 of Spiegel and Stephens (2008).
}
\keyword{datasets}
