% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostido.R
\name{Import_Triodos}
\alias{Import_Triodos}
\alias{triodos_fname}
\alias{most_recent_fname}
\alias{read_triodos_csv}
\title{Import Triodos Bank Account Transaction Data}
\usage{
triodos_fname(.date, filepath = NULL)

most_recent_fname(
  filepath = NULL,
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = triodos_fname
)

read_triodos_csv(filename)
}
\arguments{
\item{.date}{\code{Date} object, the date to be incorporated into a filename string.}

\item{filepath}{\code{character} string, the path to a folder in which to conduct the file search; default \code{NULL}.}

\item{trydate}{\code{Date} object, the most recent date within the file name from which to start the search; default
\code{Sys.Date()}.}

\item{earliest}{\code{Date} object, the earliest date within the file name beyond which the search is discontinued;
default \code{as.Date("2024-02-01")}.}

\item{fun}{\code{function}, used to incorporate \code{.date} into a filename search string; default \code{triodos_fname}.}

\item{filename}{\code{character} string, the name of a CSV file to be read.}
}
\value{
\item{\code{triodos_fname()}}{A \code{character} string representing a filename incorporating a date, of the form
\code{"Downloadyyyymmdd.csv"}, with attributes \code{"date"}, a \code{"Date"} object, and \code{"filepath"}, corresponding to the
argument of the same name (if supplied).}

\item{\code{most_recent_fname()}}{An object comprising a \code{character} string representing a filename incorporating a date
as specified in \code{fun}.}

\item{\code{read_triodos_csv()}}{CSV transaction file data formatted by Triodos Bank, as a dataframe.}
}
\description{
Functions to facilitate reading CSV transaction files downloaded from the Triodos website.

\code{triodos_fname()} returns the name of a Triodos Bank CSV format transactions file as a \code{character} string.

\code{most_recent_fname()} finds the Triodos Bank CSV format transactions file incorporating the most recent date
within its name among those in a specified folder.

\code{read_triodos_csv()} reads a CSV format transactions file downloaded from the Triodos website and returns the
contents as a data frame.
}
\details{
\code{triodos_fname()} returns a \code{character} string representing the name of a CSV format transactions file downloaded
from the Triodos website by concatenating the strings \code{"Download"}, a date of the form \code{"yyyymmdd"} and the
extension \code{".csv"} e.g., \code{"Download20240401.csv"}

\code{most_recent_fname()} searches the current folder or a folder specified using \code{filepath} for a filename
incorporating a date specified by the \code{.date} argument, typically the current date obtained using the default
\code{\link[base:Sys.time]{Sys.Date()}}. The filename incorporates the date as specified by \code{fun}, typically the default
\code{triodos_fname()} as above. If no such file exists, filenames incorporating earlier dates are searched for
successively until either a corresponding file is found or the search is discontinued upon reaching the date
specified in the \code{earliest} argument.

\code{read_triodos_csv()} reads a transactions CSV file as downloaded from the Triodos website using
\code{\link[utils:read.table]{read.csv()}} and returns the contents as a data frame. The file to be read is specified in the
\code{filename} argument and is located in the folder specified in that argument's \code{"filepath"} attribute if it has one,
otherwise in the current working directory.
}
\examples{

   dnldpath <- "~/Triodos Bank/Downloads"

   triodos_fname(Sys.Date())

   ## __________________________
   ## Current account 55545372

   triodos_fname(Sys.Date(), file.path(dnldpath, 55545372))

\dontrun{
   (file.path(dnldpath, 55545372) |>
       most_recent_fname() |>
       read_triodos_csv())
}

   ## __________________________
   ## Savings account 55596784

   triodos_fname(Sys.Date(), file.path(dnldpath, 55596784))

\dontrun{
   (savacc <- file.path(dnldpath, 55596784) |>
       most_recent_fname() |>
       read_triodos_csv())
}
}
\seealso{
\code{\link[base:as.Date]{as.Date()}}, \code{\link[utils:read.table]{read.csv()}}, \code{\link[base:Sys.time]{Sys.Date()}}.

Other Triodos: 
\code{\link{Manage_Triodos}}
}
\concept{Triodos}
\keyword{utilities}
