% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, xs = TRUE, adjust = TRUE)
}
\arguments{
\item{x}{a \code{numeric} vector.}

\item{xs}{\code{logical}, indicating whether to calculate excess kurtosis i.e., the difference from the kurtosis of the
normal distribution; default \code{TRUE}.}

\item{adjust}{\code{logical}, indicating whether \code{x} is a sample from a population; default \code{TRUE}.}
}
\value{
A \code{numeric} containing the kurtosis value.
}
\description{
Computes the kurtosis, \eqn{\gamma_{2}}{&gamma<sub>2</sub>}, of the values in \code{x} with optional adjustment to give
\eqn{G_{2}}{G<sub>2</sub>}, the expected populaton value of kurtosis from a sample.
}
\details{
Moments for samples of size \code{n} are given by: -

\deqn{m_{r} = \displaystyle \frac{\sum \left(x - \overline{x} \right)^{r}}{n}}{%
    m<sub>r</sub> = &sum;(x - x)<sup>r</sup>/n}

The (excess) kurtosis \eqn{\gamma_{2}}{&gamma;<sub>2</sub>} of a numeric variable is the fourth moment
(\eqn{m_{4}}{m<sub>4</sub>}) about the mean rendered dimensionless by dividing by the square of the second moment
(\eqn{m_{2}}{m<sub>2</sub>}), from which 3, the value of
(\eqn{m_{4}/m_{2}^2}{m<sub>4</sub>/m<sub>2</sub><sup>2</sup>}) for the normal distribution is subtracted: -

\deqn{\gamma_{2} = \displaystyle \frac{m_4}{{m_{2}}^2} - 3}{%
    &gamma;<sub>2</sub> = m<sub>4</sub> / (m<sub>2</sub>)<sup>2</sup> - 3}

The expected population value of (excess) kurtosis \eqn{G_{2}}{G<sub>2</sub>} from a sample is obtained using: -

\deqn{G_{2} = \displaystyle \frac{(n - 1)}{(n-2)(n-3)}[(n+1)\gamma_{2} + 6]}{%
    G<sub>2</sub> = ((n-1)/((n-2)(n-3)))[(n+1)&gamma;<sub>2</sub> + 6]}

(Adapted from Crawley, 2012, and Joanes and Gill, 1998.)
}
\examples{
 ## Heights of 100 randomly selected male university students, adapted from Spiegel and Stephens
 ## (Theory and Problems of Statistics. 4th edn. McGraw-Hill. 1999. ISBN 9780071755498).
 table(heights)
 hist(heights, seq(59.5, 74.5, 3))
 kurtosis(heights) 
 kurtosis(heights, adjust = FALSE)

 ## Litter sizes in albino rats (n = 815), data from King (1924; Litter production and
 ## the sex ratio in various strains of rats. The Anatomical Record 27(5), 337-366).
 table(litter_sizes)
 hist(litter_sizes, 0:12)
 kurtosis(litter_sizes) 
 kurtosis(litter_sizes, adjust = FALSE)

}
\references{
Crawley, Michael J. (2012) \emph{The R Book}. John Wiley & Sons, Incorporated. ISBN:9780470973929. p.350-352.
\href{https://onlinelibrary.wiley.com/doi/book/10.1002/9781118448908}{\doi{10.1002/9781118448908}}

Joanes, D.N., and Gill, C.A. (1998). Comparing measures of sample kurtosis and kurtosis.
\emph{Journal of the Royal Statistical Society. Series D (The Statistician)} \strong{47}(1): 183–189.
\href{https://doi.org/10.1111/1467-9884.00122}{\doi{10.1111/1467-9884.00122}}
}
\seealso{
\code{\link[stats:Distributions]{distributions}}

Other skewness: 
\code{\link{skewness}()},
\code{\link{skewness.test}()}
}
\concept{skewness}
\keyword{univar}
