% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{kurtosis.test}
\alias{kurtosis.test}
\alias{kurt.test}
\alias{stderr_kurtosis}
\title{Test of Kurtosis}
\usage{
kurtosis.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)

kurt.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)

stderr_kurtosis(n, se_method = c("Cramer", "simple"))
}
\arguments{
\item{x}{a \code{numeric} vector.}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.}

\item{se_method}{a \code{character} string specifying the method of calculating the standard error; must be one of
\code{"Cramer"} (default), or \code{"simple"}. You can specify just the initial letter.}

\item{conf.level}{the confidence level required; default \var{0.95}.}

\item{n}{an \code{integer}, the number of observations.}
}
\value{
A list with class \code{"htest"} containing the following components: -

\item{statistic}{the value of the t-statistic.}

\item{parameter}{the degrees of freedom for the t-statistic.}

\item{p.value}{the p-value for the test.}

\item{conf.int}{confidence interval of the kurtosis (95\% or other specified level).}

\item{estimate}{the estimate of kurtosis.}

\item{alternative}{a character string describing the alternative hypothesis.}

\item{method}{the character string "Kurtosis with t-test" and the standard error method used.}

\item{data.name}{a character string giving the name of the data.}
}
\description{
Computes \eqn{G_{2}}{G<sub>2</sub>}, the expected population kurtosis of the values in \var{x} using
\code{kurtosis()}, performs a \var{t}-test of its significance and calculates a confidence interval.
}
\details{
The \var{t}-statistic is given by the estimated population \code{\link{kurtosis}}, \eqn{G_{2}}{G<sub>2</sub>},
divided by its standard error, \eqn{SE_{G_{2}}}{SE<sub>G2</sub>}, where: -

\deqn{SE_{G_{2}} = \displaystyle 2(SE_{G_{1}}) \sqrt{\frac{(n^{2} - 1)}{(n-3)(n+5)}}}{%
  SE<sub>G2</sub> = 2(SE<sub>G1</sub>) &radic;((n<sup>2</sup>-1)/((n-3)(n+5)))}

(see e.g., Joanes and Gill, 1998; Wright and Herrington 2011), or alternatively its approximation,
\eqn{\sqrt (24 / n_x)}, and the associated probability is derived from the \var{t}-distribution with
\eqn{n_{x}-2}{n<sub>x</sub>-2} degrees of freedom. The \var{t}-test is conducted according to Crawley (2012),
except that the default here is a two-tailed test. The corresponding confidence interval is calculated similarly
from the quantiles of the \var{t}-distribution using both the \code{alternative} and \code{conf.level} arguments.

\code{\link[=kurt.test]{kurt.test()}} is an alias for \code{kurtosis.test()}.
}
\note{
The confidence interval is poorly described in the available literature, seems somewhat controversial and
should be used with caution.
}
\examples{
 ## Heights of 100 randomly selected male university students, adapted from Spiegel and Stephens
 ## (Theory and Problems of Statistics. 4th edn. McGraw-Hill. 1999. ISBN 9780071755498).
 table(heights)
 kurtosis.test(heights) 
 length(heights) |> stderr_kurtosis()
 kurtosis.test(heights, se_method = "simple")
 length(heights) |> stderr_kurtosis(se_method = "simple")

 ## Litter sizes in albino rats (n = 815), data from King (1924; Litter production and
 ## the sex ratio in various strains of rats. The Anatomical Record 27(5), 337-366).
 table(litter_sizes)
 kurtosis.test(litter_sizes) 
 length(litter_sizes) |> stderr_kurtosis()

}
\references{
Crawley, Michael J. (2012) \emph{The R Book}. John Wiley & Sons, Incorporated. ISBN:9780470973929. p.350-352.
\href{https://onlinelibrary.wiley.com/doi/book/10.1002/9781118448908}{\doi{10.1002/9781118448908}}

Joanes, D.N., and Gill, C.A. (1998). Comparing measures of sample skewness and kurtosis.
\emph{Journal of the Royal Statistical Society. Series D (The Statistician)} \strong{47}(1): 183–189.
\href{https://doi.org/10.1111/1467-9884.00122}{\doi{10.1111/1467-9884.00122}}

Wright, D.B., and Herrington, J.A. (2011). Problematic standard errors and confidence intervals for skewness and
kurtosis. \emph{Behavior Research Methods} \strong{43}(1): 8-17.
\href{https://doi.org/10.3758/s13428-010-0044-x}{\doi{10.3758/s13428-010-0044-x}}
}
\seealso{
\code{\link[stats:Distributions]{distributions}}, \code{\link[stats:TDist]{TDist}}, \code{\link[stats:t.test]{t.test}}

Other skewness: 
\code{\link{kurtosis}()},
\code{\link{skewness}()},
\code{\link{skewness.test}()}
}
\concept{skewness}
\keyword{htest}
