% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-cols.R
\name{kwd_cols}
\alias{kwd_cols}
\alias{kwd_cols<-}
\alias{lgl_cols}
\title{Add Logical Columns to Data Frame Flagging Presence of Keywords}
\usage{
kwd_cols(data, .look_in = Response, value)

kwd_cols(data, .look_in = Response) <- value

lgl_cols(data)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a tibble).}

\item{.look_in}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the character column in which to look for
keywords; default \code{Response}.}

\item{value}{a \code{character vector} containing the keywords to be identified.}
}
\value{
\code{kwd_cols()} returns a modified data frame with logical columns indicating the presence of keywords in
\code{.look_in}. The infix form of the function \code{kwd_cols<-()} modifies and returns the original data frame.
\code{lgl_cols()} returns a character vector of the names of all logical columns in \code{data}.
}
\description{
Function to add logical columns indicating whether or not specified keywords are present in a character column of
a data frame.
}
\details{
The character column of \code{data} identified by \code{.look_in} is searched for keywords provided in \code{value}
using \code{\link[stringr:str_detect]{str_detect}}. The keyword search is case insensitive. If no character column identified
by \code{look_in} is present in \code{data}, an error message will be given.

The logical columns resulting from use of \code{kwd_cols()} and the infix form \code{kwd_cols()<-} may be analysed
and the number of \code{TRUE} values counted using the functions \code{\link{count_lgl}} and \code{\link{sum_lgl}}.

The function \code{lgl_cols()} returns the names of all logical columns in \code{data}.
}
\examples{
(car_names <- data.frame(Response = rownames(mtcars)))

kwd <- c("Mazda", "Merc", "Toyota", "Volkswagen", "X", "450")
car_names |> kwd_cols(value = kwd)
car_names    ## Original data frame unchanged

kwd_cols(car_names)<- kwd
car_names    ## Original data frame modified by infix version

lgl_cols(car_names)
identical(kwd, lgl_cols(car_names))
rm(car_names, kwd)
}
\seealso{
\code{\link[=detective]{detective()}}, \code{\link[stringr:str_detect]{str_detect}}.

Other logical-cols: 
\code{\link{count_lgl}()},
\code{\link{list_lgl}()}
}
\concept{logical-cols}
