% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{latlon}
\alias{latlon}
\alias{latlon.default}
\alias{latlon.matrix}
\alias{latlon_dd}
\alias{print.latlon}
\title{Create Latitude and Longitude Object}
\usage{
latlon(object, ...)

\method{latlon}{default}(object, ..., decimal = FALSE, .after = c("deg", "min", "sec"))

\method{latlon}{matrix}(object, ..., decimal = FALSE, .after = c("deg", "min", "sec"))

latlon_dd(object)

\method{print}{latlon}(x, ...)
}
\arguments{
\item{object}{a \verb{numeric vector}, representing a coordinate of latitude and longitude in decimal degrees or in
degrees, minutes and seconds; or a two-column numeric \code{matrix} representing a number of such coordinates.}

\item{\dots}{further arguments passed to or from other methods.}

\item{decimal}{\code{logical}, if \code{TRUE} indicating whether the coordinate represented by \code{object} is in decimal degrees
or otherwise in degrees, minutes and seconds; default \code{FALSE}.}

\item{.after}{a \code{character} string indicating the position of the decimal point in \code{object}; must be one of
\code{"deg"} (default), \code{"min"}, or \code{"sec"}. You can specify just the initial letter.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"latlon"}, or if \code{length(object) > 1}, a list of such objects, instantiating a
coordinate of latitude and longitude in decimal degrees or degrees, minutes and seconds, comprising a list of
either two \code{"decdeg"} or two \code{"degminsec"} objects, with attribute \code{"degrtype"} indicating which of these two
types the object is.
}
\description{
The function \code{latlon()} is used to create latitude and longitude objects represented in decimal degrees or in
degrees minutes and seconds.

\code{latlon_dd()} is a convenience function such that \code{latlon_dd(...)} is equivalent to \code{latlon(..., decimal = TRUE)}.
}
\details{
\code{latlon()} is a generic S3 function. The default method works with a numeric vector of length 2 representing a
coordinate of paired latitude and longitude values. The method for class \code{"matrix"} works with a two column
numeric matrix each row of which contains paired latitude and longitude values.
}
\examples{
latlon(c(49.54621, 18.398562), decimal = TRUE)
latlon(c(49.3246368, 18.2354822))
latlon(c(493246.368, 182354.822), .after = "sec")

latlon_dd(c(49.54621, 18.398562))

(ll_mtx <- matrix(
       c(49.546210,   18.398562,
        -37.111740,  -12.288630,
        -53.104781,   73.517283,
         48.107232, -122.778671),
        ncol = 2,
        byrow = TRUE
    ))

latlon(ll_mtx, decimal = TRUE)

matrix(
       c(49.3246368,  18.2354822,
        -37.0642264, -12.1719068,
        -53.0617212,  73.3102219,
         48.0626035, -122.464322),
        ncol = 2,
        byrow = TRUE,
        dimnames = list(c("Ostravice", "Queen Mary's Peak", "Mawson Peak", "Tally Ho"))
    ) |>
latlon()

rm(ll_mtx)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()},
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
