% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litter_sizes}
\alias{litter_sizes}
\title{Rat Litter Sizes}
\format{
\subsection{\code{litter_sizes}}{

An \code{integer vector} with length \code{815}.
\describe{
\item{litter_sizes}{Sizes of albino rat litters.}
}
}
}
\source{
King, H.D. (1924). Litter production and the sex ratio in various strains of rats. \emph{The Anatomical Record},
\strong{27}(5), 337-366. \href{https://doi.org/10.1002/ar.1090270514}{\doi{10.1002/ar.1090270514}}

See also: \href{https://brownmath.com/stat/shape.htm#Example2}{https://brownmath.com}
}
\usage{
litter_sizes
}
\description{
Litter sizes in albino rats (n = 815) from Table 7 of King (1924).
}
\keyword{datasets}
