% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds-logit.r
\name{logit}
\alias{logit}
\alias{prob_from_logit}
\title{Convert Probability to Logit}
\usage{
logit(x)

prob_from_logit(x)
}
\arguments{
\item{x}{a probability from which to derive the logit, or a logit from which to derive the probability.}
}
\value{
\code{logit} returns the derived logit; \code{prob_from_logit} returns the derived probability.
}
\description{
\code{logit()} converts a probability to the corresponding logit value.

\code{prob_from_logit()} converts a logit value to the corresponding probability.
}
\details{
The \code{logit} is defined as the log of the odds or \eqn{\log(p / q)}{\%log(p / q)}, where
\eqn{q = 1 - p}{\%q = 1 - p}.

The probability may be calculated from the \code{logit} by the inverse transformation: -

\deqn{p = \exp(\displaystyle \frac{logit}{1 + logit})}{%
       p = exp(logit / (1 + logit))}
}
\examples{
logit(0.5)
prob_from_logit(0)

seq(from = 0, to = 1, by = 0.25) |> print_lf() |> logit() |> print_lf() |> prob_from_logit()

s <- seq(-10, 10, by = 2) |> print_lf() |> prob_from_logit() |> print()
s[1:6] + s[11:6]
rm(s)
}
\seealso{
\code{\link{exp}}, \code{\link{log}}, \code{\link[stats:Logistic]{plogis}} and \code{\link[stats:Logistic]{qlogis}}
and the \emph{Note} section of \code{\link[stats:Logistic]{Logistic}} in the \pkg{\link[stats]{stats}} package.

Other odds-logit: 
\code{\link{odds}()}
}
\concept{odds-logit}
\keyword{math}
