% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostido.R
\name{Manage_Rostido}
\alias{Manage_Rostido}
\alias{as_rostido}
\alias{rbind.rostido}
\alias{print.rostido}
\title{Manage Rostido Bank Account Transaction Data}
\usage{
as_rostido(data, dateformat = "\%d/\%m/\%Y")

\method{rbind}{rostido}(..., .arrange_by = NULL)

\method{print}{rostido}(x, ..., .include = NULL, maxwidth = 65L)
}
\arguments{
\item{data}{data frame, as returned by \code{read_rostido_csv()}.}

\item{dateformat}{\code{character} string, passed as the \code{format} argument to \code{\link[base:as.Date]{as.Date()}}; default
\code{"\%d/\%m/\%Y"}.}

\item{\dots}{for \code{rbind()} S3 method for class \code{"rostido"}, data frames  of class \code{"rostido"} to be combined.

for \code{print()} S3 method for class \code{"rostido"}, further arguments passed to or from other methods.}

\item{.arrange_by}{a list of expressions containing names of column(s) for sorting rows of the combined
\code{"rostido"} data frame e.g., \code{exprs(Account, Code, desc(Amount))}. Use \code{\link[dplyr:desc]{desc()}} to sort a
variable in descending order; default \code{NULL}.}

\item{x}{an object used to select a method.}

\item{.include}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> names of variables to be included or excluded when
printing a \code{"rostido"} data frame containing Rostido Bank transaction data; default \code{NULL}.}

\item{maxwidth}{an \code{integer}, maximum width for printing \code{Description} field; default \code{65L}.}
}
\value{
\item{\code{as_rostido()}}{An object of class \code{"rostido"} inheriting from \code{"data.frame"} containing reformatted Rostido
Bank transaction data.}
}
\description{
Functions to facilitate formatting and combining CSV transaction data downloaded from the Rostido website.

\code{as_rostido()} reformats a data frame containing downloaded Rostido Bank transaction data.
}
\details{
\code{as_rostido()} reformats a data frame containing Rostido Bank transaction data obtained using \code{read_rostido_csv()},
replacing \code{character} strings in the \code{Date} field with \code{"Date"} objects, and those in the \code{Amount} and \code{Balance}
fields with \code{numeric} values.

By default, if no \code{.arrange_by } argument is specified, the \code{rbind()} S3 method for class \code{"rostido"} sorts the
results by \code{Date}, \code{AccountNo} and \code{Code}.

By default, if no \code{.include} argument is specified, the \code{print()} S3 method for class \code{"rostido"} excludes the
\code{SortCode} and \code{ChequeNo} columns from the printed output.
}
\examples{

\dontrun{
   dnldpath <- "~/Rostido Bank/Downloads"

   ## __________________________
   ## Current account 55545372

   (curacc <- file.path(dnldpath, 55545372) |>
       most_recent_fname() |>
       read_rostido_csv() |>
       as_rostido())

   ## __________________________
   ## Savings account 55596784

   (savacc <- file.path(dnldpath, 55596784) |>
       most_recent_fname() |>
       read_rostido_csv() |>
       as_rostido())

   savacc |> print(.include = c(Description, Code, Amount, Balance))

   ## ______________
   ## All accounts

   rbind(curacc, savacc) ## default sort is by Date, AccountNo and Code.
   rbind(curacc, savacc, .arrange_by = exprs(AccountNo, Date, Code))
   rbind(curacc, savacc, .arrange_by = exprs(desc(Amount)))
   rbind(curacc, savacc, .arrange_by = exprs(Code, desc(Amount)))

   rm(curacc, savacc)
}
}
\seealso{
\code{\link[base:as.Date]{as.Date()}}, \code{\link[base:print]{print()}}, \code{\link[base:cbind]{rbind()}}.

Other Rostido: 
\code{\link{Import_Rostido}}
}
\concept{Rostido}
\keyword{utilities}
