% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marker}
\alias{marker}
\title{Inlinable Marker Function for Function Development}
\usage{
marker(rval = NULL, msg = "in Mimiland")
}
\arguments{
\item{rval}{value to be returned; default \code{NULL}.}

\item{msg}{ending of the output message. Since this argument follows \var{rval}, which normally would only ever
be supplied inline, its name cannot be omitted unless either it follows a comma or \code{marker()} is inlined;
default \code{"in Mimiland"}.}
}
\value{
The first argument \var{rval} if any, otherwise \code{NULL}.
}
\description{
This inlinable function outputs a user modifiable message identifying its enclosing function using
\code{\link[base]{sys.call}} then returns its first argument, if any.
}
\details{
Mainly useful for the side effect of the message, which can be used to distinguish among versions of the function under
development. Can only be used within a function and may be inlined in a \code{<}\code{\link[base:pipeOp]{piped sequence}}\code{>}
using \verb{|>}, see examples.
}
\examples{
f1 <- function() marker()
f1()

f2 <- function() marker(msg = "in BitsnBobs")
f2()

f3 <- function() marker(, "in BitsnBobs")
f3()

## Use inline in "piped" sequence
f4 <- function(x) x |> marker() |> sqrt()
f4(9)

f5 <- function(x) x |> marker(msg = "inlined in BitsnBobs") |> exp()
f5(1)

f6 <- function(x) x |> marker("inlined in BitsnBobs") |> log()
f4(1) |> f5() |> f6()

rm_objects(f, 1:6)
}
\seealso{
\code{\link[base:sys.parent]{sys.call()}}

Other utils: 
\code{\link{const}()},
\code{\link{endstop}()},
\code{\link{op-min-max}},
\code{\link{revmat}()}
}
\concept{utils}
\keyword{internal}
