% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds-logit.r
\name{odds}
\alias{odds}
\alias{prob_from_odds}
\title{Convert Probability to Odds}
\usage{
odds(x)

prob_from_odds(x)
}
\arguments{
\item{x}{a probability from which to derive the odds, or odds from which to derive the probability.}
}
\value{
\code{odds} returns the derived odds; \code{prob_from_odds} returns the derived probability.
}
\description{
\code{odds()} converts a probability to the corresponding odds.

\code{prob_from_odds()} converts odds to the corresponding proability.
}
\details{
The \code{odds} are defined as \eqn{\log(p / q)}{\%log(p / q)} where \eqn{q = 1 - p}{\%q = 1 - p}.

The probability may be calculated from the odds by the inverse transformation: -

\deqn{p = \displaystyle \frac{odds}{1 + odds}}{%
       p = odds / (1 + odds)}
}
\examples{
odds(0.5)
prob_from_odds(1)

(s <- seq(from = 0, to = 1, by = 0.125))
(s <- odds(s))
prob_from_odds(s)

rm(s) 
}
\seealso{
Other odds-logit: 
\code{\link{logit}()}
}
\concept{odds-logit}
\keyword{math}
