% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{opt_bc}
\alias{opt_bc}
\title{Optimise the Value of Lambda for a Box-Cox Transformation}
\usage{
opt_bc(x, interval = c(-5, 5), tol = 1e-10)
}
\arguments{
\item{x}{a \code{numeric} vector of data to be transformed using optimised values.}

\item{interval}{a vector containing the end-points of the interval
    to be searched for the minimum.}

\item{tol}{the desired accuracy.}
}
\value{
A numeric vector the same length as \var{x} with \code{\link[base]{attributes}} \var{lambda} and
\var{skew} representing the optimal value of \eqn{\lambda} and corresponding skewness minimum respectively.
}
\description{
Optimise the value of \eqn{\lambda} for a Box-Cox transformation by minimising skewness and return the optimally
transformed data together with the optimal \eqn{\lambda} and corresponding skewness minimum as attributes.
}
\details{
Uses the functions \code{\link{boxcox3}} to create a Box-Cox transformation function for the given dataset,
\code{\link{skewness}} to calculate the skewness of the transformed data and \code{\link[stats]{optimise}}
to search for the value of \eqn{\lambda} that minimises skewness.
}
\examples{
runif(20) |> opt_bc()
rlnorm(20) |> opt_bc()
rpois(100, 3) |> opt_bc()         ## Provokes warning
rpois(100, 3) |> opt_bc(c(-3, 5)) ## Redefining interval avoids warning
}
\seealso{
\code{\link{skewness}}

Other boxcox: 
\code{\link{boxcox3}()}
}
\concept{boxcox}
\keyword{models}
\keyword{optimize}
\keyword{regression}
