% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl-coef.R
\name{phi_coef}
\alias{phi_coef}
\title{Phi Correlation Coefficient Of Association Between Paired Binary Variables}
\usage{
phi_coef(x)
}
\arguments{
\item{x}{a square \code{matrix} containing the observations of two binary variables as a two-by-two table of counts.}
}
\value{
A numeric between -1 and 1, the \eqn{\phi} correlation coefficient.
}
\description{
The \emph{phi} correlation coefficient (or mean square contingency coefficient and denoted by \eqn{\phi} or \eqn{r\phi})
is a measure of association between two binary variables.
}
\details{
For a two-by-two contingency table \eqn{n_{11}} \eqn{n_{12}} \eqn{n_{21}} \eqn{n_{22}} the \eqn{\phi} correlation
coefficient is given by: -

\deqn{\displaystyle \phi = \frac{n_{11}n_{22} - n_{12}n_{21}}
   {\sqrt{(n_{11} + n_{21})(n_{12} + n_{22})(n_{11} + n_{12})(n_{21} + n_{22})}}}{%
   phi = (n(11).n(22) - n(12).n(21)) / \sqrt((n(11) + n(21)) * (n(12) + n(22)) * (n(11) + n(12)) * (n(21) + n(22)))}

or equivalently, the determinant of the matrix divided by the (principal) square root of the product of its four
marginal sums.
}
\examples{
## Example from Wikipedia
twobytwo <- matrix(c(6, 1, 2, 3), nrow = 2, dimnames = rep(list(c("Cat", "Dog")), 2) |>
              setNames(c("Actual", "Predicted")))
addmargins(twobytwo)

phi_coef(twobytwo)

## Example from Statology
twobytwo <- matrix(c(4, 8, 9, 4), nrow = 2, dimnames =
              list(Gender = c("Male", "Female"), Party = c("Dem", "Rep")))
addmargins(twobytwo)

phi_coef(twobytwo)

rm(twobytwo)
}
\references{
Yule, G.U. (1912). On the Methods of Measuring Association Between Two Attributes. \emph{J Royal Stat Soc}. \strong{75}
(6): 579–652. \href{https://doi.org/10.1177/10.2307/2340126}{\doi{10.2307/2340126}}.
}
\seealso{
\code{\link[base]{matrix}}, \code{\link[stats]{mcnemar.test}}

Other correl_coef: 
\code{\link{cor_coef.test}()},
\code{\link{phi_coef.test}()}
}
\concept{correl_coef}
