% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl-coef.R
\name{phi_coef.test}
\alias{phi_coef.test}
\title{Test For Association/Correlation Between Paired Binary Variables}
\usage{
phi_coef.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{a square \code{matrix} containing the observations of two binary variables as a two-by-two table of counts.}

\item{alternative}{a \code{character} string specifying the alternative hypothesis, must be one of \code{"two.sided"}
(default), \code{"greater"} or \code{"less"}.  You can specify just the initial letter.}

\item{conf.level}{numeric between 0 and 1, the confidence level required; default \var{0.95}.}
}
\value{
A list with class \code{"htest"} containing the following components: -

\item{statistic}{the value of the test statistic.}

\item{parameter}{the number of (paired) observations.}

\item{p.value}{the p-value of the test.}

\item{conf.int}{confidence interval of the \eqn{\phi} correlation coefficient (95\% or other specified level).}

\item{estimate}{the \eqn{\phi} correlation coefficient.}

\item{null.value}{the value of the association measure under the null hypothesis, always 0.}

\item{alternative}{a character string describing the alternative hypothesis.}

\item{method}{the character string "Phi correlation coefficient with confidence interval".}

\item{data.name}{a character string giving the name of the data.}
}
\description{
This function calculates a p-value for the signifinace of a \emph{phi} correlation coefficient (or mean square
contingency coefficient, \eqn{\phi} or \eqn{r\phi}), the measure of association between two binary variables, and
calculates a confidence interval.
}
\details{
The \emph{phi} coefficient is calculated using \code{\link{phi_coef}}. For derivation of the standard error and
confidence interval, see Bishop \emph{et al.} (2003), and Bonett (2021). See also \code{ci.phi()}, the \emph{confidence
interval for a phi correlation} in the
\href{https://cran.r-project.org/web/packages/statpsych/statpsych.pdf}{reference manual} for package
\CRANpkg{statpsych}.
}
\examples{
## Example from Wikipedia
twobytwo <- matrix(c(6, 1, 2, 3), nrow = 2, dimnames = rep(list(c("Cat", "Dog")), 2) |>
              setNames(c("Actual", "Predicted")))
addmargins(twobytwo)

phi_coef.test(twobytwo)
phi_coef.test(twobytwo, alternative = "less")
phi_coef.test(twobytwo, alternative = "greater")

## Example from Statology
twobytwo <- matrix(c(4, 8, 9, 4), nrow = 2, dimnames =
              list(Gender = c("Male", "Female"), Party = c("Dem", "Rep")))
addmargins(twobytwo)

phi_coef.test(twobytwo)

## Setting confidence level to 1 - p-value gives upper bound of confidence interval close to zero
pval <- phi_coef.test(twobytwo)$p.value
phi_coef.test(twobytwo, conf.level = 1 - pval)
## Similarly, with one-tailed tests setting confidence level to 1 - p-value/2 conserves the upper
## or lower CI bound with alternative = "less" or alternative = "greater" respectively
phi_coef.test(twobytwo, alternative = "less", conf.level = 1 - pval/2)
phi_coef.test(twobytwo, alternative = "greater", conf.level = 1 - pval/2)

## Example from statpsych::ci.phi(), which should return: -
##       Estimate         SE         LL        UL
## [1,] 0.1229976 0.05746271 0.01037273 0.2356224

twobytwo <- matrix(c(229, 28, 96, 24), nrow = 2, dimnames = rep(list(c("Zero", "One")), 2))
addmargins(twobytwo)

phi_coef.test(twobytwo)
## Check standard error as expected
with(phi_coef.test(twobytwo), c(stderr = estimate/statistic))

rm(twobytwo, pval)
}
\references{
Yule, G.U. (1912). On the Methods of Measuring Association Between Two Attributes. \emph{J Royal Stat Soc}.
\strong{75} (6): 579–652. \href{https://doi.org/10.1177/10.2307/2340126}{\doi{10.2307/2340126}}.

Bishop, Y.M.M., Fienberg, S.E., Holland, P.W. (1975). \emph{Discrete Multivariate Analysis.} MIT Press. (See Ch.11.)
\href{https://download.e-bookshelf.de/download/0000/0020/95/L-G-0000002095-0002339867.pdf}{ISBN 978-0-387-72805-6}.

Bonett, Douglas G. (2021).
\href{https://dgbonett.sites.ucsc.edu/statistical-methods-for-psychologists/}{Statistical Methods for Psychologists},
Volume 3: Introduction to Introduction to Categorical Data Analysis. University of California, Santa Cruz. (See 3.4
Measures of Association for 2 × 2 Tables.)
}
\seealso{
\code{\link[base]{matrix}}, \code{\link[stats]{mcnemar.test}}

Other correl_coef: 
\code{\link{cor_coef.test}()},
\code{\link{phi_coef}()}
}
\concept{correl_coef}
\keyword{htest}
