% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriever.R
\name{retriever}
\alias{retriever}
\alias{remplacer}
\title{'Function Factory' for Bespoke Data Frame Retrieval and Replacement}
\usage{
retriever(data, index, labile_data = TRUE)

remplacer(index)
}
\arguments{
\item{data}{a \code{\link[base:data.frame]{data frame}}, or a data frame extension (e.g. a \code{tibble}).}

\item{index}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the index column to be searched.}

\item{labile_data}{\code{logical}. If \code{TRUE}, \code{data} are represented in the function environment as a \code{quosure}.
If \code{FALSE}, a copy of \code{data} is saved in the function environment. Default \code{TRUE}.}
}
\value{
For \code{retriever()}, a \code{\link[base:function]{function}} for retrieval of specified columns of \code{data} in rows matching a
specified value in the \code{index} column, having the following arguments: –

\item{key}{Value to be matched in the \code{index} column.}

\item{\dots}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> character columns to return. If none are provided, all
columns are returned.}

For \code{remplacer()}, a replacement \code{function} for modifying a value in a specified column of \code{data} in rows matching a
specified value in the index column, having the following arguments: –

\item{data}{A \code{\link[base:data.frame]{data frame}}, or a data frame extension (e.g. a \code{tibble}).}

\item{key}{Value to be matched in the \code{index} column.}

\item{replace}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of the column of the value to be replaced.}

\item{value}{The replacement value of the same type as \code{replace}.}
}
\description{
\code{retriever()} creates bespoke functions for retrieval of rows of a data frame identified using a specified
index column.

\code{remplacer()} creates bespoke replacement functions for modifying rows of a data frame identified using a
specified index column.
}
\details{
The \code{key} value supplied as an argument to functions derived from both \code{retriever()} and \code{remplacer()} must be exact
matches with values in the \code{index} column for rows to be retrieved or replaced. For pattern matching, use
\code{\link[=detective]{detective()}}.

If \code{labile_data} is \code{TRUE}, \code{data} are represented in the function environment as a \code{\link[rlang:topic-quosure]{quosure}},
and functions returned by \code{retriever()} will automatically refer to the current version of \code{data} in its original
\code{\link[base:environment]{environment}}, usually the calling environment i.e., typically but not necessarily the global
environment. If \code{labile_data} is \code{FALSE}, returned functions refer to a copy of \code{data} saved in the function
environment at the time of execution of \code{retriever()}, and will not reflect any subsequent changes to the original
\code{data}.
}
\note{
The original idea was for \code{remplacer()} to create bespoke replacement functions using \code{data} captured in the function
environment as with \code{retriever()}. However, there is no immediately apparent way to do this as the object to be
"replaced" must be the first argument for replacement functions.
}
\examples{
\dontshow{starwars <- (\() dplyr::starwars)()}

## Create "retrieval" function with labile_data TRUE
retrieve_starwars <- retriever(starwars, name)

## `data` represented as "quosure" in function environment 
environment(retrieve_starwars)$data

## Retrieve selected columns for a row specified using the index
retrieve_starwars("Luke Skywalker", ends_with("color"), homeworld)

## Create "retrieval" function with labile_data FALSE
retrieve_original_starwars <- retriever(starwars, name, FALSE)

## Copy of `data` saved in function environment 
environment(retrieve_original_starwars)$data

## Retrieve selected columns for a row specified using the index
retrieve_original_starwars("Luke Skywalker", ends_with("color"), homeworld)

## Create replacement function
`replace_at_name<-` <- remplacer(name)

## Replace the value of a selected column for a row specified using the index
starwars |> replace_at_name("Luke Skywalker", homeworld) <- "Mimiland"

## Retrieve selected columns for a row specified using the index
## "retrieval" function with labile_data TRUE reflects the change
retrieve_starwars("Luke Skywalker", ends_with("color"), homeworld)

## Retrieve selected columns for a row specified using the index
## "retrieval" function with labile_data FALSE shows no change
retrieve_original_starwars("Luke Skywalker", ends_with("color"), homeworld)

rm(retrieve_starwars, retrieve_original_starwars, `replace_at_name<-`, starwars)

}
\seealso{
\code{\link[base:data.frame]{data frame}}

Other detective: 
\code{\link{detective}()},
\code{\link{wizard}()}
}
\concept{detective}
