% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostido.R
\name{rostido}
\alias{rostido}
\alias{triodos_fname}
\alias{most_recent_fname}
\alias{read_triodos_csv}
\alias{as_rostido}
\alias{rbind.rostido}
\alias{print.rostido}
\title{Manage Triodos Bank Account Transaction Data}
\usage{
triodos_fname(.date, filepath = NULL)

most_recent_fname(
  filepath = NULL,
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = triodos_fname
)

read_triodos_csv(filename)

as_rostido(data, dateformat = "\%d/\%m/\%Y")

\method{rbind}{rostido}(..., .arrange_by = NULL)

\method{print}{rostido}(x, ..., .include = NULL, maxwidth = 65L)
}
\arguments{
\item{.date}{\code{Date} object, the date to be incorporated into a filename string.}

\item{filepath}{\code{character} string, the path to a folder in which to conduct the file search; default \code{NULL}.}

\item{trydate}{\code{Date} object, the most recent date within the file name from which to start the search; default
\code{Sys.Date()}.}

\item{earliest}{\code{Date} object, the earliest date within the file name beyond which the search is discontinued;
default \code{as.Date("2024-02-01")}.}

\item{fun}{\code{function}, used to incorporate \code{.date} into a filename search string; default \code{triodos_fname}.}

\item{filename}{\code{character} string, the name of a CSV file to be read.}

\item{data}{data frame, as returned by \code{read_triodos_csv()}.}

\item{dateformat}{\code{character} string, passed as the \code{format} argument to \code{\link[base:as.Date]{as.Date()}}; default
\code{"\%d/\%m/\%Y"}.}

\item{\dots}{for \code{rbind()} S3 method for class \code{"rostido"}, data frames  of class \code{"rostido"} to be combined.

for \code{print()} S3 method for class \code{"rostido"}, further arguments passed to or from other methods.}

\item{.arrange_by}{a list of expressions containing names of column(s) for ordering rows of the combined
\code{"rostido"} data frame e.g., \code{exprs(Account, Code, desc(Amount))}. Use \code{\link[dplyr:desc]{desc()}} to sort a
variable in descending order; default \code{NULL}.}

\item{x}{an object used to select a method.}

\item{.include}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> names of variables to be included or excluded when
printing a \code{"rostido"} data frame containing Triodos Bank transaction data; default \code{NULL}.}

\item{maxwidth}{an \code{integer}, maximum width for printing \code{Description} field; default \code{65L}.}
}
\value{
\item{\code{triodos_fname()}}{A \code{character} string representing a filename incorporating a date, of the form
\code{"Downloadyyyymmdd.csv"}, with attributes \code{"date"}, a \code{"Date"} object, and \code{"filepath"}, corresponding to the
argument of the same name (if supplied).}

\item{\code{most_recent_fdate()}}{An object comprising a \code{character} string representing a filename incorporating a date
as specified in \code{fun}.}

\item{\code{read_triodos_csv()}}{CSV transaction file data formatted by Triodos Bank, as a dataframe.}

\item{\code{as_rostido()}}{An object of class \code{"rostido"} inheriting from \code{"data.frame"} containing reformatted Triodos
Bank transaction data.}
}
\description{
\code{triodos_fname()} returns the name of a Triodos Bank CSV format transactions file as a \code{character} string.

\code{most_recent_fdate()} returns the name of the Triodos Bank CSV format transactions file within a specified folder
incorporating the most recent date.

\code{read_triodos_csv()} reads a CSV format transactions file downloaded from the Triodos website and returns the
contents as a data frame.

\code{as_rostido()} reformats a data frame containing downloaded Triodos Bank transaction data.
}
\details{
These functions facilitate reading, formatting and combining CSV transaction files downloaded from the Triodos
website.

\code{triodos_fname()} returns a \code{character} string representing the name of a CSV format transactions file downloaded
from the Triodos website by concatenating the strings \code{"Download"}, a date of the form \code{"yyyymmdd"} and the
extension \code{".csv"} e.g., \code{"Download20240401.csv"}

\code{most_recent_fdate()} searches the current folder or a folder specified using \code{filepath} for a filename
incorporating a date specified by the \code{.date} argument, typically the current date obtained using the default
\code{\link[base:Sys.time]{Sys.Date()}}. The filename incorporates the date as specified by \code{fun}, typically the default
\code{triodos_fname()} as above. If no such file exists, filenames incorporating earlier dates are searched for
successively until either a corresponding file is found or the search is discontinued upon reaching the date
specified in the \code{earliest} argument.

\code{read_triodos_csv()} reads a transactions CSV file as downloaded from the Triodos website using
\code{\link[utils:read.table]{read.csv()}} and returns the contents as a data frame. The file to be read is specified in the
\code{filename} argument and is located in the folder specified in that argument's \code{"filepath"} attribute if it has one,
otherwise in the current working directory.

\code{as_rostido()} reformats a data frame containing Triodos Bank transaction data obtained using \code{read_triodos_csv()},
replacing \code{character} strings in the \code{Date} field with \code{"Date"} objects, and those in the \code{Amount} and \code{Balance}
fields with \code{numeric} values.

By default, if no \code{.arrange_by } argument is specified, the \code{rbind()} S3 method for class \code{"rostido"} orders the
results by \code{Date}, \code{AccountNo} and \code{Code}.

By default, if no \code{.include} argument is specified, the \code{print()} S3 method for class \code{"rostido"} excludes the
\code{SortCode} and \code{ChequeNo} columns from the printed output.
}
\examples{

   dnldpath <- "~/Triodos Bank/Downloads"

   triodos_fname(Sys.Date())

   ## __________________________
   ## Current account 55545372

   triodos_fname(Sys.Date(), file.path(dnldpath, 55545372))

\dontrun{
   (curacc <- file.path(dnldpath, 55545372) |>
       most_recent_fname() |>
       read_triodos_csv() |>
       as_rostido())
}

   ## __________________________
   ## Savings account 55596784

\dontrun{
   (savacc <- file.path(dnldpath, 55596784) |>
       most_recent_fname() |>
       read_triodos_csv() |>
       as_rostido())

   savacc |> print(.include = c(Description, Code, Amount, Balance))
}

   ## ______________
   ## All accounts

\dontrun{
   rbind(curacc, savacc)
   rbind(curacc, savacc, .arrange_by = exprs(desc(Amount)))
   rbind(curacc, savacc, .arrange_by = exprs(Date, Code))
   rbind(curacc, savacc, .arrange_by = exprs(Code, Amount))
   rbind(curacc, savacc, .arrange_by = exprs(Code, desc(Amount)))

   rm(curacc, savacc)
}
}
\seealso{
\code{\link[base:as.Date]{as.Date()}}, \code{\link[base:print]{print()}}, \code{\link[base:cbind]{rbind()}},
\code{\link[utils:read.table]{read.csv()}}, \code{\link[base:Sys.time]{Sys.Date()}}.
}
\keyword{utilities}
