% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{S3_Method_Debug}
\alias{S3_Method_Debug}
\alias{ls_all}
\alias{s3mag7}
\alias{browse}
\title{S3 Method Debugging Functions}
\usage{
ls_all(all.names = TRUE, env = parent.frame(), ...)

s3mag7(env = parent.frame())

browse(fn, ...)
}
\arguments{
\item{all.names}{a logical value.  If \code{TRUE}, all
    object names are returned.  If \code{FALSE}, names which begin with a
    \samp{.} are omitted.}

\item{env}{an \code{environment} to use in listing the available objects, equivalent to the \code{name} argument of
\code{\link[base:ls]{ls()}}; default \code{\link[base:sys.parent]{parent.frame()}} i.e., the \code{environment} in which the function
was called.}

\item{\dots}{additional arguments passed to \code{ls()} or \code{fn()}.}

\item{fn}{a \code{function} to be called in \dQuote{browser mode}.}
}
\value{
A \code{\link[base:character]{character vector}} for \code{ls_all()}; for \code{s3mag7()}, a named \code{\link[base:list]{list}}
containing the following elements: -

\item{\code{.Class}}{.Class is a  \code{\link[base:character]{character vector}} of \code{\link[base:class]{classes}} used to find the next
\code{\link[utils:methods]{method}}. \code{\link[base:UseMethod]{NextMethod()}} adds an attribute \code{"previous"} to .Class giving
the .Class last used for dispatch, and shifts .Class along to that used for dispatch.}

\item{\code{.Generic}}{A length-one \code{\link[base:character]{character vector}} naming the generic function for the current
\code{\link[utils:methods]{method}}.}

\item{\code{.GenericCallEnv}}{The environment of the call to be generic.}

\item{\code{.GenericDefEnv}}{The environment defining the generic, used to find \code{\link[utils:methods]{methods}} registered
for the generic.}

\item{\code{.Group}}{The generic \code{\link[base:groupGeneric]{group}} to which the \code{\link[utils:methods]{method}} belongs, if
applicable.}

\item{\code{.Method}}{A character vector (normally of length one) naming the \code{\link[utils:methods]{method}} function. (For
functions in the generic group \code{\link[base:groupGeneric]{Ops}}, it is of length two.)}

\item{\code{"object"}}{i.e., the \code{"object"} comprising the first argument of the call to \code{.Generic}.}

\item{\code{.class2("object")}}{The exact full \code{\link[base:character]{character vector}} of the classes of \code{"object"} used by
\code{\link[base:UseMethod]{UseMethod()}}.}
}
\description{
Functions intended for use in interactive mode from within the \code{\link[base:browser]{browser()}}.
}
\details{
\code{ls_all()} returns a \code{\link[base:character]{character vector}} giving the names of all objects in its caller's
\code{\link[base:environment]{environment}} including any that begin with a \sQuote{⁠.⁠}, and is convenient shorthand for
\code{ls(.all.names = TRUE)}, for instance when used from within the \code{\link[base:browser]{browser()}} in interactive mode.

Intended for use while debugging an S3 \code{\link[utils:methods]{method}} in interactive mode with the
\code{\link[base:browser]{browser()}}, \code{s3mag7()} returns a named list of seven \sQuote{special} objects in the S3 method
dispatch \code{environment}, see the \strong{Technical Details} section of \code{\link[base:UseMethod]{UseMethod}}.

\code{browse()} calls a specified function \code{fn} in "browser" mode with suitable arguments provided in \code{\dots}. The
\pkg{base} function \code{\link[base:debug]{debug}} is generally preferable.
}
\examples{
fn <- function() {
    m <- "Mimi"
    p <- "Poley"
    .b <- "Blossom"
    ls_all()
}

fn()

rm(fn)

## To run this in browser() interactive mode from R Console, select lines between
##   "## Not run:" and "## End(Not run)" and hit [shift][enter]

\dontrun{

## Two-by-two table for diagnostic test comparison
(twobytwo <- matrix(
    c(31, 12, 4, 58),
    nrow = 2, 
    dimnames = rep(list(c("+ve", "-ve")), 2) |>
        setNames(c("Test1", "Test2"))
))

browse(print_all, cohens_kappa(twobytwo))
s
where
ls_all()
s
where
ls_all()
s
where
ls_all()
s
where
ls_all()
s3mag7()
f

rm(twobytwo)
}
}
\seealso{
\code{\link[base:browser]{browser()}}, \code{\link[base:class]{class}}, \code{\link[base:debug]{debug}},
\code{\link[base:environment]{environment}}, \code{\link[utils:methods]{methods()}}, \code{\link[base:sys.parent]{parent.frame()}} and
\code{\link[base:UseMethod]{UseMethod}}

Other methods: 
\code{\link{S3Gen_Meth}},
\code{\link{method_info}()}
}
\concept{methods}
\keyword{internal}
\keyword{methods}
