% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{S3Gen_Meth}
\alias{S3Gen_Meth}
\alias{known_s3generics}
\alias{s3_in_namespace}
\alias{print.info_df}
\title{Tables of S3 Generic Functions and Methods}
\usage{
known_s3generics(.arrange_by = across(Generic:Namespace))

s3_in_namespace(namespace, .arrange_by = .data$Method)

\method{print}{info_df}(x, ...)
}
\arguments{
\item{.arrange_by}{<\code{\link[rlang:args_data_masking]{data-masked}}> names of columns or functions for ordering results
using the syntax of \pkg{\link[dplyr]{dplyr}} \code{\link[dplyr:arrange]{arrange()}}. Use \code{\link[dplyr:desc]{desc()}} to sort
by a variable in descending order.}

\item{namespace}{(unquoted) name of a \code{namespace}.}
}
\value{
A \code{\link[base:data.frame]{data.frame}} with class attributes \code{c("info_df", "catapult", "data.frame")} and columns
\code{Generic} and \code{Namespace} for \code{known_s3generics()}, and \code{Generic}, \code{Class} and \code{Method} for \code{s3_in_namespace()},
as follows: -

\item{Generic}{\code{character}, the generic function name.}

\item{Namespace}{\code{character}, the namespace environment of the generic or \code{"primitive"} in the case of
\code{\link[base:Primitive]{primitive}} (internally implemented) generics.}

\item{Class}{\code{character}, the \code{class} for the method.}

\item{Method}{\code{character}, the full method name.}
}
\description{
\code{known_s3generics()} tabulates "known" S3 generic functions.

\code{s3_in_namespace()} tabulates S3 \code{methods} in a loaded \code{namespace}.
}
\details{
\code{known_s3generics()} returns a \code{data.frame} of "known" \code{\link[base:InternalMethods]{S3 generic functions}}, based on two
\verb{character vectors} in the \pkg{\link[base]{base}} package \code{\link[base:ns-load]{namespace}}, namely
\code{\link[base:ns-internal]{.knownS3Generics}} and \code{\link[base:InternalMethods]{.S3PrimitiveGenerics}}.

\code{s3_in_namespace()} returns a \code{data.frame} of \code{\link[base:UseMethod]{S3 methods}} in a loaded
\code{\link[base:ns-load]{namespace}}.
}
\examples{
known_s3generics()
known_s3generics(.arrange_by = across(Namespace:Generic))

s3_in_namespace(BitsnBobs)
s3_in_namespace(BitsnBobs, .arrange_by = across(Class:Generic))

s3_in_namespace(utils) |> head(40)
s3_in_namespace(utils, .arrange_by = across(Class:Generic)) |> head(40)

}
\references{
R Internals: \href{https://cran.r-project.org/doc/manuals/r-devel/R-ints.html#Namespaces}{\code{1.2.2 Namespaces}}.
}
\seealso{
\code{\link[base:ns-internal]{.knownS3Generics}}, \code{\link[base:InternalMethods]{.S3PrimitiveGenerics}}
\code{\link[base:class]{class}}, \code{\link[utils:getS3method]{getS3method}}, \code{\link[base:ns-load]{loadNamespace}},
\code{\link[utils:methods]{methods}} and \code{\link[base:UseMethod]{UseMethod}}.

Other methods: 
\code{\link{S3_Method_Debug}},
\code{\link{method_info}()}
}
\concept{methods}
\keyword{internal}
\keyword{methods}
