% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{sample_size}
\alias{sample_size}
\title{Sample Size Calculaton}
\usage{
sample_size(Pexp = 0.5, d, N = NULL, conf_level = 0.95)
}
\arguments{
\item{Pexp}{\code{numeric} (or \verb{numeric vector} with all values) between 0 and 1 representing the expected prevalence;
default \var{0.5}.}

\item{d}{a positive \code{numeric} (or \verb{numeric vector} with all values) greater than zero representing the desired
absolute precision.}

\item{N}{either \code{NULL} for a large (theoretically infinite) population or a positive \code{integer} (or \verb{integer vector}
with all values) greater than zero representing the population size; default \code{NULL}.}

\item{conf_level}{the confidence level(s) required; default \var{0.95}.}
}
\value{
A \code{numeric} (or \verb{numeric vector}) giving the required sample size(s).
}
\description{
Calculate the sample size required from a population to estimate the prevalence of a dichotomous variable to a given
degree of absolute precision and with a specified level of confidence.
}
\details{
For large study populations, the sample size, \var{n}, may be calculated as: -

\deqn{n = \displaystyle \frac{z^{2}P_{exp}(1-P_{exp})}{d^{2}}}{%
       n = z^{2}Pexp(1 - Pexp) / d^{2}}

where \eqn{P_{exp}} is the expected prevalence, \var{z} is the quantile of the normal distribution corresponding
to the confidence level required and \var{d} is the desired absolute precision.

For relatively small study populations of size \var{N}, the value of \var{n} may be adjusted thus: -

\deqn{n_{adj} = \displaystyle \frac{n.N}{n + N}}{%
       nadj = n.N / (n = N)}
}
\note{
\emph{Sample Size Calculator} at \href{https://www.calculator.net}{www.calculator.net} has further useful
\href{https://www.calculator.net/sample-size-calculator.html}{information on derivation}.
}
\examples{
## Infinite population
sample_size(d = 0.05)
sample_size(d = 0.1)

## Population = 500, 750 or 1000
sample_size(d = 0.05, N = c(500L, 750L, 1000L))

## Expected prevalence = 0.125, 0.25, 0.50, 0.75 or 0.875
## Note symmetry of resulting sample sizes
sample_size(Pexp = c(0.125, 0.25, 0.50, 0.75, 0.875), d = 0.05)

## Desired absolute precision = 1\%, 5\%, 10\%, 20\%
sample_size(d = c(0.01, 0.05, 0.1, 0.2))
}
\references{
Thrusfield, M., Christley, R. In \emph{Veterinary Epidemiology} 4th Edn. John Wiley & Sons Ltd. 2018. (See Ch.13.)
\href{https://doi.org/10.1002/9781118280249}{\doi{10.1002/9781118280249}}.
}
\seealso{
Other sample-size: 
\code{\link{design_effect}()}
}
\concept{sample-size}
\keyword{survey}
