% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness.test}
\alias{skewness.test}
\alias{skew.test}
\title{Test For Skewness of a Variable}
\usage{
skewness.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)

skew.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{a \code{numeric} vector.}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.}

\item{se_method}{a character string specifying the method of calculating the standard error; must be one of
\code{"Cramer"} (default), or \code{"simple"}. You can specify just the initial letter.}

\item{conf.level}{the confidence level required; default \var{0.95}.}
}
\value{
A list with class \code{"htest"} containing the following components: -

\item{statistic}{the value of the t-statistic.}

\item{parameter}{the degrees of freedom for the t-statistic.}

\item{p.value}{the p-value for the test.}

\item{conf.int}{confidence interval of the skewness (95\% or other specified level).}

\item{estimate}{the estimate of skewness.}

\item{alternative}{a character string describing the alternative hypothesis.}

\item{method}{the character string "Skewness with t-test" and the standard error method used.}

\item{data.name}{a character string giving the name of the data.}
}
\description{
Computes \eqn{G_{1}}{G<sub>1</sub>}, the expected population skewness of the values in \var{x} using
\code{skewness()}, performs a \var{t}-test of its significance and calculates a confidence interval.
}
\details{
The \var{t}-statistic is given by the estimated population \code{\link{skewness}}, \eqn{G_{1}}{G<sub>1</sub>},
divided by its standard error, \eqn{SE_{G_{1}}}{SE<sub>G1</sub>}, where: -

\deqn{SE_{G_{1}} = \displaystyle \sqrt{\frac{6n(n - 1)}{(n-2)(n+1)(n+3)}}}{%
  SE<sub>G1</sub> = &radic;(((6n(n-1))/((n-2)(n+1)(n+3)))}

(see e.g., Joanes and Gill, 1998; Wright and Herrington 2011), or alternatively its approximation,
\eqn{\sqrt (6 / n_x)}, and the associated probability is derived from the \var{t}-distribution with
\eqn{n_{x}-2}{n<sub>x</sub>-2} degrees of freedom. The \var{t}-test is conducted according to Crawley (2012),
except that the default here is a two-tailed test. The corresponding confidence interval is calculated similarly
from the quantiles of the \var{t}-distribution using both the \code{alternative} and \code{conf.level} arguments.

\code{\link[=skew.test]{skew.test()}} is an alias for \code{skewness.test()}.
}
\note{
The confidence interval is poorly described in the available literature, seems somewhat controversial and
should be used with caution.
}
\examples{
 ## Heights of 100 randomly selected male university students, adapted from Spiegel and Stephens
 ## (Theory and Problems of Statistics. 4th edn. McGraw-Hill. 1999. ISBN 9780071755498).
 table(heights)
 skewness.test(heights) 
 skewness.test(heights, se_method = "simple")

 ## Litter sizes in albino rats (n = 815), data from King (1924; Litter production and
 ## the sex ratio in various strains of rats. The Anatomical Record 27(5), 337-366).
 table(litter_sizes)
 skewness.test(litter_sizes) 
 with(skewness.test(litter_sizes), c(stderr = estimate/statistic))

 ## Compare a range of distributions, each with the three possible alternative hypotheses
 list(
   uniform = runif(30),
   normal = rnorm(30),
   lognormal = rlnorm(30),
   poisson = rpois(30, lambda = 10),
   negbinom = rnbinom(30, mu = 4, size = 2)
 ) |>
 map(\(distrib)
     c("less", "two.sided","greater") |>
     setNames(nm = _) |>
     map(\(altern)
         with(skewness.test(distrib, altern),
             data.frame(
                 Lower = conf.int[1],
                 Upper = conf.int[2],
                 Skewness = estimate,
                 t = statistic,
                 df = parameter,
                 p = p.value,
                 sig = starsig(p.value),
                 row.names = NULL
             )
         )
     ) |>
     bind_rows(.id = "Alternative")) |>
 bind_rows(.id = "Distribution")

}
\references{
Crawley, Michael J. (2012) \emph{The R Book}. John Wiley & Sons, Incorporated. ISBN:9780470973929. p.350-352.
\href{https://onlinelibrary.wiley.com/doi/book/10.1002/9781118448908}{\doi{10.1002/9781118448908}}

Joanes, D.N., and Gill, C.A. (1998). Comparing measures of sample skewness and kurtosis.
\emph{Journal of the Royal Statistical Society. Series D (The Statistician)} \strong{47}(1): 183–189.
\href{https://doi.org/10.1111/1467-9884.00122}{\doi{10.1111/1467-9884.00122}}

Wright, D.B., and Herrington, J.A. (2011). Problematic standard errors and confidence intervals for skewness and
kurtosis. \emph{Behavior Research Methods} \strong{43}(1): 8-17.
\href{https://doi.org/10.3758/s13428-010-0044-x}{\doi{10.3758/s13428-010-0044-x}}
}
\seealso{
\code{\link[stats:Distributions]{distributions}}, \code{\link[stats:TDist]{TDist}}, \code{\link[stats:t.test]{t.test}}

Other skewness: 
\code{\link{skewness}()}
}
\concept{skewness}
\keyword{htest}
