% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars2}
\alias{starwars2}
\title{Starwars Dataset with Separated First Names and Surnames}
\format{
\subsection{\code{starwars2}}{

A data frame with 63 rows and 12 columns:
\describe{
\item{Firstname}{Firstname of the character}
\item{Surname}{Surname of the character}
\item{height}{height}
\item{mass}{mass}
\item{height}{height}
\item{hair_color}{hair color}
\item{skin_color}{skin color}
\item{eye_color}{eye color}
\item{birth_year}{year of birth}
\item{sex}{sex}
\item{gender}{gender}
\item{homeworld}{homeworld}
\item{species}{species}
}
}
}
\source{
\url{https://dplyr.tidyverse.org/reference/starwars.html}
}
\usage{
starwars2
}
\description{
\link[dplyr:starwars]{Starwars} dataset from \pkg{\link[dplyr]{dplyr}} with first names and surnames in
separate columns.
}
\keyword{datasets}
