% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars3}
\alias{starwars3}
\title{Starwars Dataset with Name and Skin Colour Only}
\format{
\subsection{\code{starwars3}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{name}{Name of the character}
\item{skin_color}{skin color}
}
}
}
\source{
\url{https://dplyr.tidyverse.org/reference/starwars.html}
}
\usage{
starwars3
}
\description{
The \code{name} and \code{skin_colour} columns from the \link[dplyr]{starwars} dataset in \pkg{\link[dplyr]{dplyr}}.
}
\keyword{datasets}
