% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BitsnBobs.R
\name{swapsign}
\alias{swapsign}
\title{Vectorised conditional sign change}
\usage{
swapsign(x, negate)
}
\arguments{
\item{x}{A numeric vector.}

\item{negate}{A logical vector of the same length as \code{x}.}
}
\value{
A numeric vector, length of \code{x}.
}
\description{
Vectorised function that changes sign of elements of a numeric vector, dependent on values of a
logical argument of the same length.
}
\examples{

swapsign(1:10, rep(c(FALSE, TRUE), 5))

}
\keyword{category}
\keyword{math}
