% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triodos.R
\name{triodos}
\alias{triodos}
\alias{file_name}
\alias{most_recent_fdate}
\alias{read_triodos}
\alias{fmt_triodos}
\title{Manage Triodos Bank Account}
\usage{
file_name(.date)

most_recent_fdate(
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = file_name
)

read_triodos(filename)

fmt_triodos(data, dateformat = "\%d/\%m/\%Y", maxwidth = 50L)
}
\arguments{
\item{.date}{\code{Date} object, date to be incorporated into a filename string.}

\item{trydate}{\code{Date} object, most recent date from which to search for file; default \code{Sys.Date()}.}

\item{earliest}{\code{Date} object, date before which search is discontinued; default \code{as.Date("2024-02-01")}.}

\item{fun}{\code{function}, used to incorporate \verb{.date' into a filename search string; default }file_name`.}

\item{filename}{\code{character} string, name of the CSV file to be read.}

\item{data}{data frame, as returned by \code{read_triodos()}.}

\item{dateformat}{\verb{character string}, passed as the \code{format} argument to \code{\link[base:as.Date]{as.Date()}}; default
\code{"\%d/\%m/\%Y"}.}

\item{maxwidth}{an \code{integer}, maximum width for printing \code{Description} field; default \code{50L}.}
}
\value{
\item{\code{file_name()}}{Filename string incorporating a date of the form \code{"Downloadyyyymmdd.csv"}.}

\item{\code{most_recent_fdate()}}{\code{"Date"} object containing the most recent date as incorporated within the specified
filename string.}

\item{\code{read_triodos()}}{CSV file data formatted by Triodos Bank, as a dataframe.}

\item{\code{fmt_triodos()}}{reformatted Triodos Bank data, as a dataframe.}
}
\description{
\code{file_name()} returns a filename string incorporating a date of the form \code{"Downloadyyyymmdd.csv"} e.g.,
\code{"Download20240401.csv"}

\code{most_recent_fdate()} returns the most recent date as incorporated within such a filename string.

\code{read_triodos()} reads a CSV file as formatted by Triodos Bank and returns the contents as a data frame.

\code{fmt_triodos()} reformats a data frame containing Triodos Bank data.
}
\details{
To follow...
}
\examples{
\dontrun{

    filepath <- "/Users/Charlie Chaplin/Triodos Bank/Downloads"
    (oldwd <- setwd(filepath))
    getwd()
    ## __________________________
    ## Current account 22245372
    setwd(paste0(filepath, "/22245372"))
    getwd()

    most_recent_fdate() |>
        file_name() |>
        read_triodos() |>
        fmt_triodos() |>
        _[, -2]

    setwd(oldwd)

} ## end \dontrun
}
\keyword{utilities}
