% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{waypoint}
\alias{waypoint}
\title{Waypoint}
\usage{
waypoint(..., .fmt = c("decdeg", "degmin", "degminsec"))
}
\arguments{
\item{...}{two \code{numeric} values or two \code{"coord"} objects.}

\item{.fmt}{\code{character} string indicating the desired format; must be one of \code{"decdeg"}
(default), \code{"degmin"} or \code{"degminsec"}.}
}
\value{
An object of \code{class} \code{"waypoint"} comprising a named list of two \code{"coord"} objects representing
the latitude and longitude values of a waypoint.
}
\description{
A robust representation of a geographic or GPS waypoint comprising latitude and langitude values
implemented as \code{\link[=coord]{"coord"}} objects.
}
\details{
Waypoints may be defined using (i) decimal degrees, (ii) degrees and minutes, or (iii) degrees,
minutes and seconds.
...
...
}
\examples{
## Decimal degrees
waypoint(coord(51.50776), coord(-0.127924))
waypoint(as_coord(51.50776), as_coord(-0.127924))
waypoint(51.50776, -0.127924)

## Degrees amd minutes
waypoint(coord(51L, 30.4659), coord(-0L, 07.6754))
waypoint(as_coord(5130.4659, .fmt = "degmin"), as_coord(-007.6754, .fmt = "degmin"))
waypoint(5130.4659, -007.6754, .fmt = "degmin")

## Degrees, minutes and seconds
waypoint(coord(51L, 30L, 27.95), coord(-0L, 07L, 40.53))
waypoint(as_coord(5130.4659, .fmt = "degminsec"), as_coord(-00740.53, .fmt = "degminsec"))
waypoint(5130.4659, -00740.53, .fmt = "degminsec")
}
\seealso{
Other coord: 
\code{\link{coord}()}
}
\concept{coord}
