% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective.R
\name{wizard}
\alias{wizard}
\alias{data_wizard}
\title{Extract, Sort Unique Values, and Paste Column from Data Frame}
\usage{
wizard(data, col, .collapse = NULL)

data_wizard(data)
}
\arguments{
\item{data}{a data frame, or a data frame extension (e.g. a tibble).}

\item{col}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name of character column to extract.}

\item{.collapse}{an optional character string to separate the results, see \code{\link[base:paste]{paste}}; default
\code{NULL}.}
}
\value{
A vector of the same type as \code{col} or a single character string if a value for \code{.collapse} is
supplied.
}
\description{
\code{wizard()} extracts and sorts unique values of a selected column from a data frame, and optionally pastes into
a character string.

\code{data_wizard()} does the same across all columns in a  data frame.
}
\details{
\code{wizard()} can be useful within a piped sequence to quickly extract or review \code{\link[base:sort]{sorted}},
\code{\link[base:unique]{unique}} contents of a column and optionally collapse into a single character string using
\code{\link[base:paste]{paste}} by providing a suitable value for \code{.collapse}.
}
\examples{
\dontshow{starwars <- dplyr::starwars}

starwars |> wizard(homeworld)
starwars |> wizard(homeworld, ", ")
starwars |> wizard(homeworld, "\t") |> cat()
starwars |> wizard(homeworld, "\n") |> cat()

data_wizard(mtcars) 

\dontshow{rm(starwars)}
}
\seealso{
\code{\link[base:paste]{paste}}, \code{\link[base:sort]{sort}}, \code{\link[base:unique]{unique}}.

Other detective: 
\code{\link{detective}()},
\code{\link{retriever}()}
}
\concept{detective}
