% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{const}
\alias{const}
\title{Make a Constant with Active Binding}
\usage{
const(name, value)
}
\arguments{
\item{name}{a symbol, the name to be assigned to the constant.}

\item{value}{any valid R object, including a function}
}
\description{
A constant is an object the value of which cannot be altered once assigned.
}
\details{
See reference.
}
\examples{
const(bar, "irish")
bar
try(bar <- "bavarian")

}
\references{
\href{https://iqis.netlify.app/post/2019/07/22/how-to-make-a-constant-in-r/}{
Siqi Zhang, 2019: Make a Constant in R with Active Binding}.
}
\seealso{
\code{\link[base:bindenv]{lockBinding()}}

Other utils: 
\code{\link{endstop}()},
\code{\link{marker}()},
\code{\link{op-min-max}},
\code{\link{revmat}()}
}
\concept{utils}
\keyword{utilities}
