% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{Geographic or GPS Coordinate}
\usage{
coord(deg = 0, min = NULL, sec = NULL, .latorlon = c(NA, "lat", "lon"))
}
\arguments{
\item{deg}{\code{numeric}, representing the number of degrees. Must be of type \code{integer} if \code{min} or
\code{sec} are provided, otherwise type \code{double}; default \code{0}.}

\item{min}{\code{numeric}, representing the number of minutes. If \code{sec} provided, must be of type
\code{integer}, otherwise \code{double}; default \code{NULL}.}

\item{sec}{\code{double}, representing the number of seconds; default \code{NULL}.}

\item{.latorlon}{a \code{character} string, either \code{"lat"} or \code{"lon"} indicating whether the
coordinate represented is of latitude or longitude, or \code{NA} (the default).}
}
\value{
An object of class \code{"coord"} instantiating a coordinate. See \emph{Details}.
}
\description{
Geographic or GPS coordinate class
}
\details{
\code{coord()} creates a robust representation of a geographic or GPS cordinate based on the value of
\code{deg}, \code{min} and \code{sec} instatiated as an objects of class \code{"coord"}. Objects of \code{"coord"} class
contain a \code{list} with one, two or three \code{numeric} values named \code{"deg"}, \code{"min"}, \code{"sec"}, depending
on whether the cordinate in question is represented in decimal degrees, in (integer) degrees and
(decimal) minutes, or else in (integer) degrees, (integer) minutes, and (decimal) seconds.

The value provided in argument \code{dec} should have a decimal point after the number of whole
degrees in the case of decimal degrees. Likewise the value provided in argument \code{min} should have
a decimal point, after the number of whole minutes in the case of degrees and minutes, and the
value provided in argument \code{sec} after the number of whole seconds in the case of degrees,
minutes and seconds. In all other cases, arguments \code{deg} and \code{min} must be of type \code{integer}.

\code{"coord"} objects have \code{character} attribute \code{latorlon}, which may be \code{"lat"} for latitude,
\code{"lon"} for longitude or \code{NA}, and a \code{logical} attribute \code{"negative"}, which when \code{TRUE}
signifies a negative coordinate i.e., S or W rather than N or E.

The total value in degrees, minutes and seconds may not be greater than \verb{180˚}, while the
minutes and seconds components (if present) must be less than  \verb{60˚}. If latitude is
represented, (i.e., \code{latorlon} attribute is \code{"lat"}), its  maximum absolute value is \verb{90˚}.
}
\examples{
## Decimal degrees
coord()

coord(51.507765)
coord(-51.507765)
coord(51.507765,,, "lat")
coord(-51.507765,,, "lat")
coord(51.507765,,, "lon")
coord(-51.507765,,, "lon")

## Degrees and (decimal) minutes
coord(51L, 30.4659)
coord(-51L, 30.4659)
coord(51L, 30.4659,, "lat")
coord(-51L, 30.4659,, "lat")
coord(51L, 30.4659,, "lon")
coord(-51L, 30.4659,, "lon")

## Degrees, minutes and (decimal) seconds
coord(51L, 30L, 27.95)
coord(-51L, 30L, 27.95)
coord(51L, 30L, 27.95, "lat")
coord(-51L, 30L, 27.95, "lat")
coord(51L, 30L, 27.95, "lon")
coord(-51L, 30L, 27.95, "lon")
}
\seealso{
Other coord: 
\code{\link{as_degminsec}()}
}
\concept{coord}
