% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{decdeg}
\alias{decdeg}
\alias{decdeg.default}
\alias{print.decdeg}
\title{Create Decimal Degrees Object}
\usage{
decdeg(object, ...)

\method{decdeg}{default}(object, ...)

\method{print}{decdeg}(x, ...)
}
\arguments{
\item{object}{numeric, representing a coordinate of latitude or longitude in decimal degrees.}

\item{\dots}{further arguments passed to or from other methods.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"decdeg"}, representing a coordinate of latitude or longitude in decimal degrees
represented by a numeric of type \code{double} with maximum absolute value of 180.
}
\description{
The function \code{decdeg()} is used to create (latitude or longitude) coordinate objects represented in decimal degrees.
}
\details{
\code{decdeg()} is a generic S3 function. The default method works with a numeric (\code{double}) representing a
coordinate of latitude or longitude in decimal degrees.
}
\examples{
decdeg(49.54621)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()},
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
