% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{decdeg}
\alias{decdeg}
\alias{decdeg.default}
\alias{print.decdeg}
\title{Create Decimal Degrees Object}
\usage{
decdeg(object, ...)

\method{decdeg}{default}(object, ..., .latorlon = c(NA, "lat", "lon"))

\method{print}{decdeg}(x, ...)
}
\arguments{
\item{object}{numeric, representing one or more coordinates of latitude or longitude in decimal degrees.}

\item{\dots}{further arguments passed to or from other methods.}

\item{.latorlon}{a character string indicating whether the coordinate represented by \code{object} is a latitude or
longitude; must be one of \code{NA} (default), \code{"lat"}, or \code{"lon"}. You can specify just the initial letter.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"decdeg"}, or if \code{length(object) > 1}, a \code{list} of such objects, instantiating a
coordinate of latitude or longitude in decimal degrees represented by a numeric of type \code{double} with maximum
absolute value of 180 degrees.
}
\description{
The function \code{decdeg()} is used to create (latitude or longitude) coordinate objects represented in decimal degrees.
}
\details{
\code{decdeg()} is a generic S3 function. The default method works with a numeric (\code{double}) vector representing one
or more coordinates of latitude or longitude in decimal degrees.
}
\examples{
decdeg(49.54621)
decdeg(c(lat = 49.54621, lon = 18.2354822))
decdeg(c(lat = -37.11174, lon = -12.28863))
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()},
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
