% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{degminsec}
\alias{degminsec}
\alias{as_degminsec}
\alias{as_degminsec.default}
\alias{as_degminsec.decdeg}
\alias{as_degminsec.list}
\alias{as_degminsec.latlon}
\alias{print.degminsec}
\title{Create Degrees, Minutes and Seconds Object}
\usage{
degminsec(
  object,
  .fmt = c("deg", "min", "sec"),
  .latorlon = c(NA, "lat", "lon")
)

as_degminsec(object, ...)

\method{as_degminsec}{default}(object, ..., .latorlon = c(NA, "lat", "lon"))

\method{as_degminsec}{decdeg}(object, ...)

\method{as_degminsec}{list}(object, ...)

\method{as_degminsec}{latlon}(object, ...)

\method{print}{degminsec}(x, ...)
}
\arguments{
\item{object}{\code{numeric}, representing one or more coordinates of latitude or longitude in degrees, minutes and
seconds.}

\item{.fmt}{a \code{character} string indicating the position of the decimal point in \code{object}; must be one of
\code{"deg"} (default), \code{"min"}, or \code{"sec"}. You can specify just the initial letter.}

\item{.latorlon}{a \code{character} string indicating whether the coordinate represented by \code{object} is a latitude or
longitude; must be one of \code{NA} (default), \code{"lat"}, or \code{"lon"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"degminsec"}, or if \code{length(object) > 1}, a list of such objects, representing a
coordinate of latitude or longitude in degrees, minutes and seconds as a named list with components: -

\item{deg}{degrees represented by an integer with maximum absolute value of 180.}

\item{min}{minutes represented by a positive integer with value less than 60.}

\item{sec}{seconds represented by a positive numeric with value less than 60.}

Attribute \code{"negative"} indicates whether \code{object} was originally a negative number i.e. if \code{TRUE}, the
value represents a west or south rather than north or east coordinate. Attribute \code{".latorlon"} indicates
whether the object is a coordinate of latitude or longitude.
}
\description{
The function \code{degminsec()} is used to create (latitude or longitude) coordinate objects represented in degrees,
minutes and seconds.

\code{as_degminsec()} converts decimal degrees to coordinate objects represented in  degrees, minutes and seconds.
}
\details{
\code{degminsec()} works with a numeric (\code{double}) representing one or more coordinates of latitude or longitude in
degrees, minutes and seconds. The argument \code{.fmt} indicates the position of the decimal point, which may be placed
after the degrees, the minutes or the (whole) seconds, and by default assumed to be placed after the degrees.

\code{as_degminsec()} is an S3 function that works with individual coordinates supplied as \verb{numeric values} or as
\code{\link[=decdeg]{"decdeg"}} objects, or with lists of such coordinates.  It also accepts, latitude and
longitude values paired in a \code{\link[=latlon]{"latlon"}} object (see examples), or lists of \code{"latlon"}
paired coordinates.
}
\examples{
degminsec(49.3246368)
degminsec(4932.46368, .fmt = "min")
degminsec(493246.368, .fmt = "sec")

degminsec(49.3246368, .latorlon = "lat")
degminsec(18.2354822, .latorlon = "lon")

degminsec(-37.0642264, .latorlon = "lat")
degminsec(-12.1719068, .latorlon = "lon")

degminsec(c(lat = 49.3246368, lon = 18.2354822))
degminsec(c(lat = -370642.264, lon = -121719.068), .fmt = "sec")

as_degminsec(49.54621, .latorlon = "lat")

(coord <- decdeg(49.54621, .latorlon = "lat"))
as_degminsec(coord)

(coords <- latlon_dd(c(49.54621, 18.398562)))
as_degminsec(coords)

rm(coord, coords)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{latlon}()}
}
\concept{degreeconvert}
\keyword{utilities}
