% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{degminsec}
\alias{degminsec}
\alias{degminsec.default}
\alias{print.degminsec}
\title{Create Degrees, Minutes and Seconds Object}
\usage{
degminsec(x, ...)

\method{degminsec}{default}(x, ..., pointafter = c("deg", "min", "sec"))

\method{print}{degminsec}(x, ...)
}
\arguments{
\item{x}{numeric, representing a coordinate of latitude or longitude in degrees, minutes and seconds.}

\item{pointafter}{a character string specifying the position of the decimal point in \code{x}; must be one of
"deg" (default), "min", or "sec". You can specify just the initial letter.}
}
\value{
An object of class \code{"degminsec"}, representing a coordinate of latitude or longitude in degrees, minutes
and seconds as a named list with components: -

\item{dd}{degrees represented by an integer with maximum absolute value of 180.}

\item{mm}{minutes represented by a positive integer with value less than 60.}

\item{ss}{seconds represented by a positiveinteger with value less than 60.}
}
\description{
The function \code{degminsec()} is used to create (latitude or longitude) coordinate objects represented in degrees,
minutes and seconds.
}
\details{
\code{degminsec()} is a generic S3 function. The default method works with a numeric (\code{double}) representing a
coordinate of latitude or longitude in degrees, minutes and seconds. The argument \code{pointafter} indicates the
position of the decimal point, which may be placed after the degrees, the minutes or the (whole) seconds, and by
default assumed to be placed after the degrees.
}
\examples{
degminsec(49.32464)
degminsec(49.32464, pointafter = "deg")
degminsec(4932.464, pointafter = "min")
degminsec(493246.4, pointafter = "sec")
}
\seealso{
Other degreeconvert: 
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
