% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{dms_to_decdeg}
\alias{dms_to_decdeg}
\alias{dms_to_decdeg.default}
\title{Convert Degrees, Minutes and Seconds to Decimal Degrees}
\usage{
dms_to_decdeg(x, ...)

\method{dms_to_decdeg}{default}(dms)
}
\arguments{
\item{dms}{an object of class \code{\link[=degminsec]{"degminsec"}}, representing a coordinate of latitude or longitude in
degrees, minutes and seconds or a list of "degminsec" objects.}
}
\value{
numeric, a coordinate of latitude or longitude in decimal degrees
}
\description{
Convert degrees, minutes and seconds to decimal degrees.
}
\details{
\code{dms_to_decdeg()} is a vectorised function that works with individual coordinates, latitude and longitude
values paired in a named list (see examples) or with a longer list of coordinates.
}
\examples{
degminsec(49.32464) |> dms_to_decdeg()

(coords <- list(lat = degminsec(49.32464), long = degminsec(18.2354822)))
dms_to_decdeg(coords)

rm(coords)
}
\seealso{
Other degreeconvert: 
\code{\link{degminsec}()}
}
\concept{degreeconvert}
\keyword{utilities}
