% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{dms_to_decdeg}
\alias{dms_to_decdeg}
\alias{dms_to_decdeg.degminsec}
\alias{dms_to_decdeg.default}
\alias{dms_to_decdeg.list}
\title{Convert Degrees, Minutes and Seconds to Decimal Degrees}
\usage{
dms_to_decdeg(object, ...)

\method{dms_to_decdeg}{degminsec}(object, ...)

\method{dms_to_decdeg}{default}(object, ..., pointafter = c("deg", "min", "sec"))

\method{dms_to_decdeg}{list}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[=degminsec]{"degminsec"}}, representing a coordinate of latitude or longitude in
degrees, minutes and seconds or a list of "degminsec" objects.}

\item{...}{further arguments passed to or from other methods.}

\item{pointafter}{a character string specifying the position of the decimal point in \code{x}; must be one of
"deg" (default), "min", or "sec". You can specify just the initial letter.}
}
\value{
An object of class \code{\link[=decdeg]{"decdeg"}}, representing a coordinate of latitude or longitude in decimal
degrees represented by a numeric of type \code{double} with maximum absolute value of 180, or a list of such objects.
}
\description{
Convert degrees, minutes and seconds to decimal degrees.
}
\details{
\code{dms_to_decdeg()} is an S3 function that works with individual coordinates, latitude and longitude
values paired in a named list (see examples) or with a longer list of coordinates.
}
\examples{

dms_to_decdeg(49.32464)
dms_to_decdeg(49.32464, pointafter = "deg")
dms_to_decdeg(4932.464, pointafter = "min")
dms_to_decdeg(493246.4, pointafter = "sec")

(coord <- degminsec(49.32464))
dms_to_decdeg(coord)

(coords <- list(lat = degminsec(49.32464), long = degminsec(18.2354822)))
dms_to_decdeg(coords)

rm(coord, coords)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{degminsec}()}
}
\concept{degreeconvert}
\keyword{utilities}
