% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{latlon}
\alias{latlon}
\alias{latlon.default}
\alias{latlon_dd}
\alias{latlon_dms}
\alias{print.latlon}
\title{Create Latitude and Longitude Object}
\usage{
latlon(object, ...)

\method{latlon}{default}(object, ..., decimal = FALSE, .after = c("deg", "min", "sec"))

latlon_dd(object)

latlon_dms(object, .after = c("deg", "min", "sec"))

\method{print}{latlon}(x, ...)
}
\arguments{
\item{object}{a numeric vector, representing a pair of coordinates of latitude and longitude in decimal degrees
or in degrees, minutes and seconds.}

\item{\dots}{further arguments passed to or from other methods.}

\item{decimal}{logical, if \code{TRUE} indicating whether the coordinate represented by \code{object} is in decimal degrees
or otherwise in degrees, minutes and seconds; default \code{FALSE}.}

\item{.after}{a character string indicating the position of the decimal point in \code{object}; must be one of
\code{"deg"} (default), \code{"min"}, or \code{"sec"}. You can specify just the initial letter.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"latlon"}, or if \code{length(object) > 1}, a list of such objects, instantiating a
coordinate of latitude and longitude in decimal degrees or degrees, minutes and seconds, comprising a list of
either two \code{"decdeg"} or two \code{"degminsec"} objects, with attribute \code{"coordtype"} indicating which of these two
types the object is.
}
\description{
The function \code{latlon()} is used to create latitude and longitude objects represented in decimal degrees or in
degrees minutes and seconds.
}
\details{
\code{latlon()} is a generic S3 function. ...
}
\examples{
latlon(c(49.54621, 18.398562), decimal = TRUE)
latlon(c(49.3246368, 18.2354822))

latlon_dd(c(49.54621, 18.398562))
latlon_dms(c(49.3246368, 18.2354822))
latlon_dms(c(493246.368, 182354.822), .after = "sec")
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()},
\code{\link{dms_to_decdeg}()}
}
\concept{degreeconvert}
\keyword{utilities}
