% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeconvert.R
\name{latlon}
\alias{latlon}
\alias{latlon.default}
\alias{latlon.matrix}
\alias{latlon_dd}
\alias{print.latlon}
\title{Create Latitude and Longitude Object}
\usage{
latlon(object, ...)

\method{latlon}{default}(object, ..., decimal = FALSE, .after = c("deg", "min", "sec"))

\method{latlon}{matrix}(object, ..., decimal = FALSE, .after = c("deg", "min", "sec"))

latlon_dd(object)

\method{print}{latlon}(x, ...)
}
\arguments{
\item{object}{a \verb{numeric vector}, representing a coordinate of latitude and longitude in decimal degrees or in
degrees, minutes and seconds; or a two-column numeric \code{matrix} representing a number of such coordinates.}

\item{\dots}{further arguments passed to or from other methods.}

\item{decimal}{\code{logical}, if \code{TRUE} indicating whether the coordinate represented by \code{object} is in decimal degrees
or otherwise in degrees, minutes and seconds; default \code{FALSE}.}

\item{.after}{a \code{character} string indicating the position of the decimal point in \code{object}; must be one of
\code{"deg"} (default), \code{"min"}, or \code{"sec"}. You can specify just the initial letter.}

\item{x}{object to be printed.}
}
\value{
An object of class \code{"latlon"} instantiating a coordinate of latitude and longitude in decimal degrees or
degrees, minutes and seconds, comprising a list of either two \code{"decdeg"} or two \code{"degminsec"} objects, with
attribute \code{"degrtype"} indicating which of these two types the object is; or if \code{length(object) > 1}, a list
of \code{"latlon"} objects.
}
\description{
The function \code{latlon()} is used to create latitude and longitude objects representing paired coordinates in either
decimal degrees or degrees, minutes and seconds.

\code{latlon_dd()} is a convenience function, such that \code{latlon_dd(...)} is equivalent to \code{latlon(..., decimal = TRUE)}.
}
\details{
\code{latlon()} is a generic S3 function. The default method works with a numeric vector of length 2 representing a
coordinate of paired latitude and longitude values. The method for class \code{"matrix"} works with a two column
numeric matrix each row of which contains paired latitude and longitude values.
}
\examples{
## Decimal degrees
latlon(c(49.54621, 18.398562), decimal = TRUE)

## degrees, minutes and seconds
latlon(c(49.3246368, 18.2354822))
latlon(c(493246.368, 182354.822), .after = "sec")

## Decimal degrees—convenience function
latlon_dd(c(49.54621, 18.398562))

## Two-column numeric `matrix` in decimal degrees
(ll_mtx <- matrix(
       c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781,
         -0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283),
        ncol = 2,
        dimnames = list(
            c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument",
              "Null Island", "Tristan da Cunha", "Mawson Peak")
        )
    ))

latlon(ll_mtx, decimal = TRUE)

## Two-column numeric `matrix` in degrees, minutes and seconds
## Deg Min Sec
ll_mtx[1:14] <- c(
    51.3027954, 49.3246368, 48.0626035, 38.5322178, 0, -37.0642264, -53.0617212,
    -0.0740526, 18.2354822, -122.464322, -77.0206871, 0, -12.1719068, 73.3102219
)

ll_mtx

latlon(ll_mtx)

rm(ll_mtx)
}
\seealso{
Other degreeconvert: 
\code{\link{decdeg}()},
\code{\link{decdeg_to_dms}()},
\code{\link{degminsec}()}
}
\concept{degreeconvert}
\keyword{utilities}
