% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-dplyr.R, R/reexport-ggplot2.R,
%   R/reexport-purrr.R, R/reexport-rlang.R, R/reexport-tibble.R,
%   R/reexport-tidyselect.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{across}
\alias{bind_rows}
\alias{desc}
\alias{element_text}
\alias{theme_get}
\alias{theme_set}
\alias{theme_update}
\alias{map}
\alias{map2_chr}
\alias{map_chr}
\alias{map_dbl}
\alias{map_int}
\alias{eval_tidy}
\alias{set_names}
\alias{tibble}
\alias{all_of}
\alias{any_of}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{last_col}
\alias{matches}
\alias{num_range}
\alias{starts_with}
\alias{where}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{across}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{desc}}}

  \item{ggplot2}{\code{\link[ggplot2:element]{element_text}}, \code{\link[ggplot2]{theme_get}}, \code{\link[ggplot2:theme_get]{theme_set}}, \code{\link[ggplot2:theme_get]{theme_update}}}

  \item{purrr}{\code{\link[purrr]{map}}, \code{\link[purrr:map]{map_chr}}, \code{\link[purrr:map]{map_dbl}}, \code{\link[purrr:map]{map_int}}, \code{\link[purrr:map2]{map2_chr}}}

  \item{rlang}{\code{\link[rlang]{eval_tidy}}, \code{\link[rlang]{set_names}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{where}}}
}}

