% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rostido.R
\name{rostido}
\alias{rostido}
\alias{file_name}
\alias{most_recent_fdate}
\alias{read_triodos_csv}
\alias{as_rostido}
\alias{rbind.rostido}
\alias{print.rostido}
\title{Manage Triodos Bank Account Transaction Data}
\usage{
file_name(.date)

most_recent_fdate(
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = file_name
)

read_triodos_csv(filename)

as_rostido(data, dateformat = "\%d/\%m/\%Y")

\method{rbind}{rostido}(..., .arrange_by = across(Date:Code))

\method{print}{rostido}(x, ..., .include = NULL, maxwidth = 65L)
}
\arguments{
\item{.date}{\code{Date} object, the date to be incorporated into a filename string.}

\item{trydate}{\code{Date} object, the most recent date from which to search for file; default \code{Sys.Date()}.}

\item{earliest}{\code{Date} object, the earliest date to search for within the file name, beyond which the search is
discontinued; default \code{as.Date("2024-02-01")}.}

\item{fun}{\code{function}, used to incorporate \code{.date} into a filename search string; default \code{file_name}.}

\item{filename}{\code{character} string, the name of a CSV file to be read.}

\item{data}{data frame, as returned by \code{read_triodos_csv()}.}

\item{dateformat}{\verb{character string}, passed as the \code{format} argument to \code{\link[base:as.Date]{as.Date()}}; default
\code{"\%d/\%m/\%Y"}.}

\item{\dots}{for \code{rbind()} S3 method for class \code{"rostido"}, data frames  of class \code{"rostido"} to be combined.

for \code{print()} S3 method for class \code{"rostido"}, further arguments passed to or from other methods.}

\item{.arrange_by}{<\code{\link[rlang:args_data_masking]{data-masking}}> quoted name(s) of column(s) for ordering
results. Use \code{\link[dplyr:desc]{desc()}} to sort by variables in descending order; default \code{across(Date:Code)}.}

\item{x}{an object used to select a method.}

\item{.include}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> names of variables to be included or excluded when
printing a \code{"rostido"} data frame containing Triodos Bank transaction data; default \code{NULL}.}

\item{maxwidth}{an \code{integer}, maximum width for printing \code{Description} field; default \code{65L}.}
}
\value{
\item{\code{file_name()}}{Filename string incorporating a date of the form \code{"Downloadyyyymmdd.csv"}.}

\item{\code{most_recent_fdate()}}{\code{"Date"} object containing the most recent date as incorporated within the specified
filename string.}

\item{\code{read_triodos_csv()}}{CSV transaction file data formatted by Triodos Bank, as a dataframe.}

\item{\code{as_rostido()}}{An object of class \code{"rostido"} inheriting from \code{"data.frame"} containing reformatted Triodos
Bank transaction data.}
}
\description{
\code{file_name()} returns a character string representing the name of a CSV format transactions file downloaded from
the Triodos bank website.

\code{most_recent_fdate()} returns the most recent date incorporated within such a filename in the current folder.

\code{read_triodos_csv()} reads a Triodos Bank transactions CSV file and returns the contents as a
data frame.

\code{as_rostido()} reformats a data frame containing downloaded Triodos Bank transaction data.
}
\details{
These four functions facilitate reading and formatting CSV transaction files downloaded from the Triodos bank
website.

\code{file_name()} returns a character string representing the name of a transactions file downloaded from the Triodos
bank website in CSV format i.e., concatenating the strings \code{"Download"}, a date of the form \code{"yyyymmdd"} and the
extension \code{".csv"} e.g., \code{"Download20240401.csv"}

\code{most_recent_fdate()} searches the current folder for a filename as returned by \code{file_name()}, typically
incorporating the current date obtained using the default \code{.date} argument \code{\link[base:Sys.time]{Sys.Date()}}. If no
such file exists, the previous dates are used successively until a corresponding file is found, the search being
discontinued on reaching the date specified in the \code{earliest} argument.

\code{read_triodos_csv()} reads a transactions CSV file downloaded from the Triodos bank website and returns the
contents as a data frame.

\code{as_rostido()} reformats a data frame containing Triodos Bank transaction data obtained using \code{read_triodos_csv()},
replacing \code{character} strings in the \code{Date} field with \code{"Date"} objects, and in the \code{Amount} and \code{Balance} fields
with \code{numeric} values.

By default, if no \code{.include} argument is specified, the \code{print()} S3 method for class \code{"rostido"} excludes the
\code{SortCode} and \code{ChequeNo} columns from the printed output.
}
\examples{
\dontrun{

   filepath <- "~/Triodos Bank/Downloads"
   (oldwd <- setwd(filepath))
   getwd()

   ## __________________________
   ## Current account 55545372
   setwd(paste0(filepath, "/55545372"))
   getwd()

   (curacc <- most_recent_fdate() |>
       file_name() |>
       read_triodos_csv() |>
       as_rostido())

   ## __________________________
   ## Savings account 55596784

   setwd(paste0(filepath, "/55596784"))
   getwd()

   (savacc <- most_recent_fdate() |>
        file_name() |>
        read_triodos_csv() |>
        as_rostido())

   savacc |> print(.include = c(Description, Code, Amount, Balance))

   ## ______________
   ## All accounts
   rbind(curacc, savacc)

   rm(curacc, savacc)
   setwd(oldwd)

}
}
\seealso{
\code{\link[base:print]{print()}}, \code{\link[base:cbind]{rbind()}}.
}
\keyword{utilities}
