% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triodos.R
\name{triodos}
\alias{triodos}
\alias{file_name}
\alias{most_recent_fdate}
\alias{read_triodos}
\alias{fmt_triodos}
\title{Manage Triodos Bank Account}
\usage{
file_name(.date)

most_recent_fdate(
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = file_name
)

read_triodos(filename)

fmt_triodos(data, dateformat = "\%d/\%m/\%Y", maxwidth = 50L)
}
\arguments{
\item{.date}{\code{Date} object, date to be incorporated into a filename string.}

\item{trydate}{\code{Date} object, most recent date from which to search for file; default \code{Sys.Date()}.}

\item{earliest}{\code{Date} object, date before which search is discontinued; default \code{as.Date("2024-02-01")}.}

\item{fun}{\code{function}, used to incorporate \verb{.date' into a filename search string; default }file_name`.}

\item{filename}{\code{character} string, name of the CSV file to be read.}

\item{data}{data frame, as returned by \code{read_triodos()}.}

\item{dateformat}{\verb{character string}, passed as the \code{format} argument to \code{\link[base:as.Date]{as.Date()}}; default
\code{"\%d/\%m/\%Y"}.}

\item{maxwidth}{an \code{integer}, maximum width for printing \code{Description} field; default \code{50L}.}
}
\value{
\item{\code{file_name()}}{Filename string incorporating a date of the form \code{"Downloadyyyymmdd.csv"}.}

\item{\code{most_recent_fdate()}}{\code{"Date"} object containing the most recent date as incorporated within the specified
filename string.}

\item{\code{read_triodos()}}{CSV file data formatted by Triodos Bank, as a dataframe.}

\item{\code{fmt_triodos()}}{reformatted Triodos Bank data, as a dataframe.}
}
\description{
\code{file_name()} returns a character string representing the name of a transactions file downloaded from the Triodos
bank website in CSV format.

\code{most_recent_fdate()} returns the most recent date as incorporated within such a filename string.

\code{read_triodos()} reads a CSV file as formatted by Triodos Bank and returns the contents as a data frame.

\code{fmt_triodos()} reformats a data frame containing Triodos Bank data.
}
\details{
These four functions facilitate reading and formatting CSV transaction files downloaded from the Triodos bank
website.

\code{file_name()} returns a character string representing the name of a transactions file downloaded from the Triodos
bank website in CSV format i.e., concatenating the strings \code{"Download"}, a date of the form \code{"yyyymmdd"} and the
extension \code{".csv"} e.g., \code{"Download20240401.csv"}

\code{most_recent_fdate()} searches the current folder for a filename as returned by \code{file_name()}, typically
incorporating the current date obtained using the default \code{.date} argument \code{\link[base:Sys.time]{Sys.Date()}}. If no
such file exists, the previous dates are used successively until a corresponding file is found, the search being
discontinued on reaching the date specified in the \code{earliest} argument.

\code{read_triodos()} reads a file downloaded from the Triodos bank website in CSV format and returns the contents as a
data frame.

\code{fmt_triodos()} reformats a data frame containing Triodos Bank data obtained using \code{read_triodos()}, replacing
\code{character} strings in the \code{Date} field with \code{"Date"} objects, and in the \code{Amount} and \code{Balance} fields with
\code{numeric} values. It also truncates \code{character} strings in the \code{Description} field to the length specified in the
\code{maxwidth} argument and saves the full version of any \code{Description} exceeding \code{maxwidth} as \code{attributes}
\code{"descr_no"} and \code{"descr"}.
}
\examples{
\dontrun{

    filepath <- "/Users/Charlie Chaplin/Triodos Bank/Downloads"
    (oldwd <- setwd(filepath))
    getwd()
    ## __________________________
    ## Current account 22245372
    setwd(paste0(filepath, "/22245372"))
    getwd()

    most_recent_fdate() |>
        file_name() |>
        read_triodos() |>
        fmt_triodos() |>
        _[, -2]

    setwd(oldwd)

}
}
\keyword{utilities}
