% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{waypoint}
\alias{waypoint}
\title{Waypoint}
\usage{
waypoint(..., .fmt = c("decdeg", "degmin", "degminsec"))
}
\arguments{
\item{...}{two \code{numeric} values or two \code{"coord"} objects.}

\item{.fmt}{\code{character} string indicating the desired format; must be one of \code{"decdeg"}
(default), \code{"degmin"} or \code{"degminsec"}.}
}
\value{
An object of \code{class} \code{"waypoint"} comprising a named list of two \code{"coord"} objects representing
the latitude and longitude values of a waypoint.
}
\description{
A robust representation of a geographic or GPS waypoint comprising latitude and langitude values
implemented as \code{\link[=coord]{"coord"}} objects.
}
\details{
Waypoints may be defined using (i) decimal degrees, (ii) degrees and minutes, or (iii) degrees,
minutes and seconds. They may easily be converted between formats using \code{\link[=as_coord]{as_coord()}}.
}
\examples{
## Create "waypoint" objects

## Decimal degrees
(wp_dd <- waypoint(coord(51.507765), coord(-0.127924)))
waypoint(as_coord(51.507765), as_coord(-0.127924))
waypoint(51.507765, -0.127924)

## Degrees amd minutes
(wp_dm <- waypoint(coord(51L, 30.4659), coord(, -07.6754)))
waypoint(as_coord(5130.4659, .fmt = "degmin"), as_coord(-007.6754, .fmt = "degmin"))
waypoint(5130.4659, -007.6754, .fmt = "degmin")

## Degrees, minutes and seconds
(wp_dms <- waypoint(coord(51L, 30L, 27.95), coord(, -07L, 40.53)))
waypoint(as_coord(513027.95, .fmt = "degminsec"), as_coord(-00740.53, .fmt = "degminsec"))
waypoint(513027.95, -00740.53, .fmt = "degminsec")

## Convert between "waypoint" object formats

## To decimal degrees
wp_dd |> as_coord(.fmt = "decdeg")
wp_dm |> as_coord(.fmt = "decdeg")
wp_dms |> as_coord(.fmt = "decdeg")

## To degrees and minutes
wp_dd |> as_coord(.fmt = "degmin")
wp_dm |> as_coord(.fmt = "degmin")
wp_dms |> as_coord(.fmt = "degmin")

## To degrees, minutes and seconds
wp_dd |> as_coord(.fmt = "degminsec")
wp_dm |> as_coord(.fmt = "degminsec")
wp_dms |> as_coord(.fmt = "degminsec")

rm(wp_dd, wp_dm, wp_dms)
}
\seealso{
Other coord: 
\code{\link{coord}()}
}
\concept{coord}
